### 2022-05-21 #4412 @Kevin-Mattheus-Moerman
@edejong-caltech can you check that potentially invalid DOI :point_up: ? You can call `@editorialbot check references` to check them again.
