### 2022-05-24 505c63d
Remove invalid DOI reference
