+++ paper/paper.bib
@@ -272,14 +272,6 @@
   doi = {10.5194/gmd-9-1455-2016}
 }
 
-@article{Howell_1949,
-  title = {The growth of cloud drops in uniformly cooled air},
-  author = {Howell, W.E.},
-  doi = {10.1175/1520-0469(1949)006%3C0134:TGOCDI%3E2.0.CO;2},
-  year = {1949},
-  journal = {J. Atmos. Sci.}
-}
-
 @article{Lange_1978,
   title = {ADPIC—A Three-Dimensional Particle-in-Cell Model for the Dispersal of Atmospheric Pollutants and its Comparison to Regional Tracer Studies},
   author = {Lange, R.},
