### 2022-05-29 f0902f5
bump Numba requirement
