### 2022-05-30 #4412 @Kevin-Mattheus-Moerman
Note to handling editor, this is a second paper, after their first JOSS publication https://joss.theoj.org/papers/10.21105/joss.03219
