### 2022-05-31 8d487e5
an option to fix thd in condensation dynamics is added
