### 2022-06-01 88b2f2b
changing boolean name to update_thd
