### 2022-06-01 0d7972f
test fixed thd in condensation for Shipway and Hill example
