+++ tests/smoke_tests/shipway_and_hill_2012/test_few_steps.py
@@ -28,6 +28,7 @@ def test_few_steps_no_precip(particle_reservoir_depth, plot=False):
     )
     settings.particle_reservoir_depth = particle_reservoir_depth
     settings.t_max = 50 * settings.dt
+    settings.condensation_update_thd = True
     simulation = Simulation(settings)
 
     # Act
@@ -80,3 +81,38 @@ def test_few_steps_no_precip(particle_reservoir_depth, plot=False):
 
     assert max(mean_profile_over_last_steps("ripening rate")) > 0
     assert max(mean_profile_over_last_steps("deactivating rate")) > 0
+
+
+def test_fixed_thd():
+    # Arrange
+    n_sd_per_gridbox = 128
+    settings = Settings(
+        n_sd_per_gridbox=n_sd_per_gridbox,
+        dt=30 * si.s,
+        dz=60 * si.m,
+        precip=False,
+        rho_times_w_1=2 * si.m / si.s * si.kg / si.m**3,
+    )
+    settings.t_max = 50 * settings.dt
+    settings.condensation_update_thd = False
+    simulation = Simulation(settings)
+
+    # Act
+    output = simulation.run().products
+
+    # Assert
+    mean_profile_over_last_steps = lambda var: np.mean(
+        output[var][output["z"] >= 0, -10:], axis=1
+    )
+    sd_prof = mean_profile_over_last_steps("super droplet count per gridbox")
+    assert 0.5 * n_sd_per_gridbox < min(sd_prof) < 1.5 * n_sd_per_gridbox
+    assert 0.5 * n_sd_per_gridbox < max(sd_prof) < 1.5 * n_sd_per_gridbox
+
+    assert 0.01 < max(mean_profile_over_last_steps("peak supersaturation")) < 0.1
+    assert min(mean_profile_over_last_steps("ql")) < 1e-10
+    assert 0.6 < max(mean_profile_over_last_steps("ql")) < 0.9
+    assert max(mean_profile_over_last_steps("activating rate")) == 0
+    assert max(mean_profile_over_last_steps("ripening rate")) == 0
+    assert max(mean_profile_over_last_steps("deactivating rate")) == 0
+
+    assert sum(np.amin(output["thd"], axis=1)) == sum(np.amax(output["thd"], axis=1))
