+++ tests/smoke_tests/shipway_and_hill_2012/test_1d_exporters.py
@@ -1,3 +1,4 @@
+# pylint: disable = redefined-outer-name
 from tempfile import NamedTemporaryFile, TemporaryDirectory
 
 import numpy as np
