### 2022-06-03 d6b57dc
Write common fragmentation fn limiter
