+++ PySDM/backends/impl_numba/methods/collisions_methods.py
@@ -355,6 +355,19 @@ class CollisionsMethods(BackendMethods):
 
     @staticmethod
     @numba.njit(**{**conf.JIT_FLAGS})
+    def __fragmentation_limiters(n_fragment, frag_size, v_max, vmin, nfmax, x_plus_y):
+        for i in numba.prange(len(n_fragment)):  # pylint: disable=not-an-iterable
+            n_fragment[i] = x_plus_y[i] / frag_size[i]
+            if frag_size[i] > v_max[i]:
+                n_fragment[i] = 1
+            elif frag_size[i] < vmin:
+                n_fragment[i] = 1
+
+            if nfmax is not None:
+                n_fragment[i] = min(n_fragment[i], nfmax)
+
+    @staticmethod
+    @numba.njit(**{**conf.JIT_FLAGS})
     def __slams_fragmentation_body(n_fragment, probs, rand):
         for i in numba.prange(len(n_fragment)):  # pylint: disable=not-an-iterable
             probs[i] = 0.0
@@ -370,33 +383,26 @@ class CollisionsMethods(BackendMethods):
 
     @staticmethod
     @numba.njit(**{**conf.JIT_FLAGS})
-    def __exp_fragmentation_body(
-        *, n_fragment, scale, frag_size, v_max, x_plus_y, rand, vmin, nfmax
-    ):
+    def __exp_fragmentation_body(*, scale, frag_size, rand):
         """
         Exponential PDF
         """
-        for i in numba.prange(len(n_fragment)):  # pylint: disable=not-an-iterable
+        for i in numba.prange(len(frag_size)):  # pylint: disable=not-an-iterable
             frag_size[i] = -scale * np.log(1 - rand[i])
-            if frag_size[i] > v_max[i]:
-                n_fragment[i] = 1
-            elif frag_size[i] < vmin:
-                n_fragment[i] = 1
-            else:
-                n_fragment[i] = x_plus_y[i] / frag_size[i]
-            if nfmax is not None:
-                n_fragment[i] = min(n_fragment[i], nfmax)
 
     def exp_fragmentation(
         self, *, n_fragment, scale, frag_size, v_max, x_plus_y, rand, vmin, nfmax
     ):
         self.__exp_fragmentation_body(
-            n_fragment=n_fragment.data,
             scale=scale,
+            frag_size=frag_size.data,
+            rand=rand.data,
+        )
+        self.__fragmentation_limiters(
+            n_fragment=n_fragment.data,
             frag_size=frag_size.data,
             v_max=v_max.data,
             x_plus_y=x_plus_y.data,
-            rand=rand.data,
             vmin=vmin,
             nfmax=nfmax,
         )
@@ -404,7 +410,7 @@ class CollisionsMethods(BackendMethods):
     @staticmethod
     @numba.njit(**{**conf.JIT_FLAGS})
     def __feingold1988_fragmentation_body(
-        *, n_fragment, scale, frag_size, v_max, x_plus_y, rand, fragtol, vmin, nfmax
+        *, n_fragment, scale, frag_size, x_plus_y, rand, fragtol
     ):
         """
         Scaled exponential PDF
@@ -412,14 +418,6 @@ class CollisionsMethods(BackendMethods):
         for i in numba.prange(len(n_fragment)):  # pylint: disable=not-an-iterable
             log_arg = max(1 - rand[i] * scale / x_plus_y[i], fragtol)
             frag_size[i] = -scale * np.log(log_arg)
-            if frag_size[i] > v_max[i]:
-                n_fragment[i] = 1
-            elif frag_size[i] < vmin:
-                n_fragment[i] = 1
-            else:
-                n_fragment[i] = x_plus_y[i] / frag_size[i]
[…]
