### 2022-06-03 aa4a4a9
readVTK_1d moved to PySDM-examples.util: test changed accordingely - reason: vtk is missing on Windows 32bits
