+++ PySDM/exporters/__init__.py
@@ -4,4 +4,4 @@ Exporters handling output to metadata-rich file formats incl. netCDF and VTK
 from .netcdf_exporter import NetCDFExporter
 from .netcdf_exporter_1d import NetCDFExporter_1d, readNetCDF_1d
 from .vtk_exporter import VTKExporter
-from .vtk_exporter_1d import VTKExporter_1d, readVTK_1d
+from .vtk_exporter_1d import VTKExporter_1d
