+++ PySDM/exporters/vtk_exporter_1d.py
@@ -5,13 +5,7 @@ vtk exporter of particle attributes for 1d simulations
 import os
 
 import numpy as np
-import vtk
 from pyevtk.hl import pointsToVTK
-
-# pylint: disable = import-error, no-name-in-module
-from vtk.util import numpy_support as VN
-
-# pylint: enable = import-error, no-name-in-module
 
 
 class VTKExporter_1d:
@@ -71,25 +65,3 @@ class VTKExporter_1d:
     def run(self):
         for time_index_and_value in enumerate(self.settings.save_spec_and_attr_times):
             self._export_attributes(time_index_and_value)
-
-
-def readVTK_1d(file):
-
-    reader = vtk.vtkXMLUnstructuredGridReader()  # pylint: disable = no-member
-    reader.SetFileName(file)
-    reader.Update()
-
-    vtk_output = reader.GetOutput()
-
-    z = np.zeros(vtk_output.GetNumberOfPoints())
-    for i in range(vtk_output.GetNumberOfPoints()):
-        z[i] = vtk_output.GetPoint(i)[2]
-
-    data = {}
-    data["z"] = z
-    for i in range(vtk_output.GetPointData().GetNumberOfArrays()):
-        data[vtk_output.GetPointData().GetArrayName(i)] = VN.vtk_to_numpy(
-            vtk_output.GetPointData().GetArray(i)
-        )
-
-    return data
