### 2022-06-06 d40cdfe
addressing pylint
