### 2022-06-09 cac4f32
add tmate block to test workflow to debug segfaults interactively
