+++ .github/workflows/main.yml
@@ -49,3 +49,7 @@ jobs:
       - if: matrix.platform != 'ubuntu-latest'
         run: pytest -p no:unraisableexception -We
 
+      - if: ${{ failure() }}
+        uses: mxschmitt/action-tmate@v3
+        with:
+          limit-access-to-actor: true
