### 2022-06-10 0c5bfab
try adding GC collect() calls in between tests
