### 2022-06-10 4750e88
remove explicit gc calls
