+++ tests/__init__.py
@@ -1,7 +1,1 @@
-import gc
-import pytest
 
-
-@pytest.fixture(autouse=True)
-def ensure_gc():
-    gc.collect()
