### 2022-06-10 71b4c7d
Fix output size in fragmentation test
