### 2022-06-11 b9e4578
add # pylint: disable=too-many-arguments for __fragmentation_limiters()
