### 2022-06-15 70a99db
add a start point for PyPartMC-importing tests
