+++ tests/__init__.py
@@ -1,0 +1,8 @@
+import gc
+
+import pytest
+
+
+@pytest.fixture(autouse=True)
+def ensure_gc():
+    gc.collect()
