+++ tests/__init__.py
@@ -1,8 +1,0 @@
-import gc
-
-import pytest
-
-
-@pytest.fixture(autouse=True)
-def ensure_gc():
-    gc.collect()
