+++ .github/workflows/tests+pypi.yml
@@ -1,4 +1,4 @@
-name: PySDM
+name: tests+pypi
 
 defaults:
   run:
@@ -11,6 +11,8 @@ on:
     branches: [ main ]
   schedule:
     - cron: '0 13 * * 4'
+  release:
+    types: [published]
 
 jobs:
   build:
@@ -53,3 +55,33 @@ jobs:
 #         uses: mxschmitt/action-tmate@v3
 #         with:
 #           limit-access-to-actor: true
+
+  dist:
+    runs-on: ubuntu-latest
+    needs: [build]
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          fetch-depth: 0  # https://github.com/pypa/setuptools_scm/issues/480
+      - uses: actions/setup-python@v2
+        with:
+          python-version: "3.10"
+
+      - run: pip install twine build
+
+      - run: |
+          python -m build 2>&1 | tee build.log
+          exit `fgrep -i warning build.log | wc -l`
+
+      - run: twine check --strict dist/*
+
+      - if: github.event_name == 'push' && github.ref == 'refs/heads/main'
+        uses: pypa/gh-action-pypi-publish@master
+        with:
+          password: ${{ secrets.TEST_PYPI_API_TOKEN }}
+          repository_url: https://test.pypi.org/legacy/
+
+      - if: startsWith(github.ref, 'refs/tags')
+        uses: pypa/gh-action-pypi-publish@master
+        with:
+          password: ${{ secrets.PYPI_API_TOKEN }}
