### 2022-06-18 a0840f5
addressing pylint suggestion
