### 2022-06-18 7466e8e
fix README badge and warning checks
