+++ PySDM/backends/impl_numba/methods/collisions_methods.py
@@ -61,6 +61,7 @@ def break_up(  # pylint: disable=too-many-arguments
     max_multiplicity,
     breakup_rate,
     breakup_rate_deficit,
+    warn_overflows,
 ):  # pylint: disable=too-many-branches
     gamma_tmp = 0
     gamma_deficit = gamma[i]
@@ -106,7 +107,7 @@ def break_up(  # pylint: disable=too-many-arguments
                 attributes[a, k] /= tmp2
                 attributes[a, j] = attributes[a, k]
 
-        if overflow_flag:
+        if overflow_flag and warn_overflows:
             warn("overflow", __file__)
             break
 
@@ -278,6 +279,7 @@ class CollisionsMethods(BackendMethods):
         breakup_rate_deficit,
         is_first_in_pair,
         max_multiplicity,
+        warn_overflows,
     ):
         # pylint: disable=not-an-iterable,too-many-nested-blocks
         for i in numba.prange(length // 2):
@@ -312,6 +314,7 @@ class CollisionsMethods(BackendMethods):
                     max_multiplicity,
                     breakup_rate,
                     breakup_rate_deficit,
+                    warn_overflows,
                 )
             flag_zero_multiplicity(j, k, multiplicity, healthy)
 
@@ -332,6 +335,7 @@ class CollisionsMethods(BackendMethods):
         breakup_rate,
         breakup_rate_deficit,
         is_first_in_pair,
+        warn_overflows,
     ):
         max_multiplicity = np.iinfo(multiplicity.data.dtype).max // 2e5
         self.__collision_coalescence_breakup_body(
@@ -351,6 +355,7 @@ class CollisionsMethods(BackendMethods):
             breakup_rate_deficit=breakup_rate_deficit.data,
             is_first_in_pair=is_first_in_pair.indicator.data,
             max_multiplicity=max_multiplicity,
+            warn_overflows=warn_overflows,
         )
 
     @staticmethod
