+++ PySDM/particulator.py
@@ -143,6 +143,7 @@ class Particulator:  # pylint: disable=too-many-public-methods
         breakup_rate,
         breakup_rate_deficit,
         is_first_in_pair,
+        warn_overflows,
     ):
         idx = self.attributes._ParticleAttributes__idx
         healthy = self.attributes._ParticleAttributes__healthy_memory
@@ -165,6 +166,7 @@ class Particulator:  # pylint: disable=too-many-public-methods
                 breakup_rate=breakup_rate,
                 breakup_rate_deficit=breakup_rate_deficit,
                 is_first_in_pair=is_first_in_pair,
+                warn_overflows=warn_overflows,
             )
         else:
             self.backend.collision_coalescence(
