+++ tests/smoke_tests/partmc/test_partmc.py
@@ -1,11 +1,6 @@
 # pylint: disable=missing-module-docstring,missing-class-docstring,missing-function-docstring
 
-import sys
-
-import pytest
-
-if sys.platform == "linux":
-    import PyPartMC
+import PyPartMC
 
 
 def test_partmc():
