### 2022-07-07 a73f961
enable RDP login on Appveyor to debug platform_machine issue on 32 bit Windows
