+++ appveyor.yml
@@ -41,3 +41,6 @@ test_script:
   - ps: |
       iex "$env:PYTHON -m pytest -We -p no:unraisableexception"
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
+
+on_finish:
+  - ps: $blockRdp = $true; iex ((new-object net.webclient).DownloadString('https://raw.githubusercontent.com/appveyor/ci/master/scripts/enable-rdp.ps1'))
