### 2022-07-07 a8f6405
Add complex breakup testing positivity check
