### 2022-07-08 a978eaa
Change thd_tol to RH_tol
