+++ PySDM/backends/impl_thrust_rtc/methods/condensation_methods.py
@@ -325,7 +325,7 @@ class CondensationMethods(ThrustRTCBackendMethods):
         kappa,
         f_org,
         rtol_x,
-        rtol_thd,
+        rtol_RH,
         timestep,
         counters,
         cell_order,
