+++ PySDM/particulator.py
@@ -101,7 +101,7 @@ class Particulator:  # pylint: disable=too-many-public-methods
             RH=self.environment.get_predicted("RH"),
         )
 
-    def condensation(self, *, rtol_x, rtol_thd, counters, RH_max, success, cell_order):
+    def condensation(self, *, rtol_x, rtol_RH, counters, RH_max, success, cell_order):
         self.backend.condensation(
             solver=self.condensation_solver,
             n_cell=self.mesh.n_cell,
@@ -113,6 +113,7 @@ class Particulator:  # pylint: disable=too-many-public-methods
             rhod=self.environment["rhod"],
             thd=self.environment["thd"],
             qv=self.environment["qv"],
+            RH=self.environment["RH"],
             dv=self.environment.dv,
             prhod=self.environment.get_predicted("rhod"),
             pthd=self.environment.get_predicted("thd"),
@@ -120,7 +121,7 @@ class Particulator:  # pylint: disable=too-many-public-methods
             kappa=self.attributes["kappa"],
             f_org=self.attributes["dry volume organic fraction"],
             rtol_x=rtol_x,
-            rtol_thd=rtol_thd,
+            rtol_RH=rtol_RH,
             v_cr=self.attributes["critical volume"],
             timestep=self.dt,
             counters=counters,
