+++ PySDM/backends/impl_numba/methods/condensation_methods.py
@@ -29,6 +29,7 @@ class CondensationMethods(BackendMethods):
         rhod,
         thd,
         qv,
+        RH,
         dv,
         prhod,
         pthd,
@@ -36,7 +37,7 @@ class CondensationMethods(BackendMethods):
         kappa,
         f_org,
         rtol_x,
-        rtol_thd,
+        rtol_RH,
         timestep,
         counters,
         cell_order,
@@ -58,6 +59,7 @@ class CondensationMethods(BackendMethods):
             rhod=rhod.data,
             thd=thd.data,
             qv=qv.data,
+            RH=RH.data,
             dv_mean=dv,
             prhod=prhod.data,
             pthd=pthd.data,
@@ -65,7 +67,7 @@ class CondensationMethods(BackendMethods):
             kappa=kappa.data,
             f_org=f_org.data,
             rtol_x=rtol_x,
-            rtol_thd=rtol_thd,
+            rtol_RH=rtol_RH,
             timestep=timestep,
             counter_n_substeps=counters["n_substeps"].data,
             counter_n_activating=counters["n_activating"].data,
@@ -92,6 +94,7 @@ class CondensationMethods(BackendMethods):
         rhod,
         thd,
         qv,
+        RH,
         dv_mean,
         prhod,
         pthd,
@@ -99,7 +102,7 @@ class CondensationMethods(BackendMethods):
         kappa,
         f_org,
         rtol_x,
-        rtol_thd,
+        rtol_RH,
         timestep,
         counter_n_substeps,
         counter_n_activating,
@@ -143,12 +146,13 @@ class CondensationMethods(BackendMethods):
                     f_org,
                     thd[cell_id],
                     qv[cell_id],
+                    RH[cell_id],
                     dthd_dt,
                     dqv_dt,
                     md,
                     rhod_mean,
                     rtol_x,
-                    rtol_thd,
+                    rtol_RH,
                     timestep,
                     counter_n_substeps[cell_id],
                 )
@@ -175,7 +179,7 @@ class CondensationMethods(BackendMethods):
         n_substeps_min = math.ceil(timestep / dt_range[1])
 
         @numba.njit(**jit_flags)
-        def adapt_substeps(args, n_substeps, thd, rtol_thd):
+        def adapt_substeps(args, n_substeps, thd, RH, rtol_RH):
             n_substeps = np.maximum(n_substeps_min, n_substeps // multiplier)
             success = False
             for burnout in range(fuse + 1):
@@ -189,24 +193,26 @@ class CondensationMethods(BackendMethods):
                         ),
                         return_value=(0, False),
                     )
-                thd_new_long, success = step_fake(args, timestep, n_substeps)
+                thd_new_long, RH_new_long, success = step_fake(args, timestep, n_substeps)
                 if success:
                     break
                 n_substeps *= multiplier
             for burnout in range(fuse + 1):
                 if burnout == fuse:
                     return warn("burnout (short)", __file__, return_value=(0, False))
-                thd_new_short, success = step_fake(
+                thd_new_short, RH_new_short, success = step_fake(
                     args, timestep, n_substeps * multiplier
                 )
                 if not success:
                     return warn("short failed", __file__, return_value=(0, False))
-                dthd_long = thd_new_long - thd
-                dthd_short = thd_new_short - thd
-                error_estimate = np.abs(dthd_long - multiplier * dthd_short)
-                thd_new_long = thd_new_short
-                if within_tolerance(error_estimate, thd, rtol_thd):
+
+                dRH_long = RH_new_long - RH
+                dRH_short = RH_new_short - RH
+                error_estimate = np.abs(dRH_long - multiplier * dRH_short)
+                RH_new_long = RH_new_short
[…]
