### 2022-07-08 d1f15a2
Change thd_tol to RH_tol
