+++ PySDM/backends/impl_thrust_rtc/methods/condensation_methods.py
@@ -339,7 +339,7 @@ class CondensationMethods(
         kappa,
         f_org,
         rtol_x,
-        rtol_thd,
+        rtol_RH,
         timestep,
         counters,
         cell_order,

[…]
