### 2022-07-08 0ab37de
disable PyPartMC test on 32 bit machines
