### 2022-07-08 ed27b71
print platform.machine() in appveyor logs
