### 2022-07-08 9c6f1ef
disable installing PyPartMC on 32 bit Python through appveyor ps script
