+++ appveyor.yml
@@ -43,5 +43,5 @@ test_script:
       iex "$env:PYTHON -m pytest -We -p no:unraisableexception"
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
 
-on_finish:
-  - ps: $blockRdp = $true; iex ((new-object net.webclient).DownloadString('https://raw.githubusercontent.com/appveyor/ci/master/scripts/enable-rdp.ps1'))
+#on_finish:
+#  - ps: $blockRdp = $true; iex ((new-object net.webclient).DownloadString('https://raw.githubusercontent.com/appveyor/ci/master/scripts/enable-rdp.ps1'))
