### 2022-07-08 88ac2a5
making pylint & black happy; bumping test-time-req for examples
