+++ tests/smoke_tests/parcel/abdul_razzak_ghan_2000/test_single_supersaturation_peak.py
@@ -27,12 +27,7 @@ from PySDM.physics import si
 )
 @pytest.mark.parametrize("rtol_x", (1e-7,))
 @pytest.mark.parametrize("scheme", ("PySDM",))
-def test_single_supersaturation_peak(
-        scheme,
-        rtol_x,
-        rtol_RH,
-        plot=False
-):
+def test_single_supersaturation_peak(scheme, rtol_x, rtol_RH, plot=False):
     # arrange
     products = (
         PySDM_products.WaterMixingRatio(unit="g/kg", name="ql"),
@@ -58,13 +53,7 @@ def test_single_supersaturation_peak(
     builder = Builder(backend=CPU(), n_sd=n_sd)
     builder.set_environment(env)
     builder.add_dynamic(AmbientThermodynamics())
-    builder.add_dynamic(
-        Condensation(
-            adaptive=True,
-            rtol_x=rtol_x,
-            rtol_RH=rtol_RH
-        )
-    )
+    builder.add_dynamic(Condensation(adaptive=True, rtol_x=rtol_x, rtol_RH=rtol_RH))
 
     r_dry, concentration = ConstantMultiplicity(spectrum).sample(n_sd)
     v_dry = builder.formulae.trivia.volume(radius=r_dry)
