### 2022-07-08 df1d9fb
making pylint & black happy; bumping test-time-req for examples
