### 2022-07-08 8f04c69
making pylint & black happy; bumping test-time-req for examples
