### 2022-07-08 6c90dff
fix GPU backend condensation signature
