+++ PySDM/backends/impl_thrust_rtc/methods/condensation_methods.py
@@ -332,6 +332,7 @@ class CondensationMethods(
         rhod,
         thd,
         qv,
+        RH,
         dv,
         prhod,
         pthd,
