### 2022-07-08 dde3ab5
fix GPU backend condensation signature
