### 2022-07-08 9b73291
fix GPU backend condensation signature
