+++ PySDM/backends/impl_thrust_rtc/methods/condensation_methods.py
@@ -320,6 +320,7 @@ class CondensationMethods(
         rhod,
         thd,
         qv,
+        RH,
         dv,
         prhod,
         pthd,
