+++ PySDM/backends/impl_numba/test_helpers/scipy_ode_condensation_solver.py
@@ -42,6 +42,7 @@ def _condensation(
         rhod=particulator.environment["rhod"].data,
         thd=particulator.environment["thd"].data,
         qv=particulator.environment["qv"].data,
+        RH=particulator.environment["RH"].data,
         dv_mean=particulator.environment.dv,
         prhod=particulator.environment.get_predicted("rhod").data,
         pthd=particulator.environment.get_predicted("thd").data,
@@ -178,14 +179,15 @@ def _make_solve(formulae):  # pylint: disable=too-many-statements,too-many-local
         thd,
         qv,
         rhod,
+        __,
         dthd_dt,
         dqv_dt,
         drhod_dt,
         m_d_mean,
-        __,
         ___,
-        dt,
         ____,
+        dt,
+        _____,
     ):
         n_sd_in_cell = len(cell_idx)
         y0 = np.empty(n_sd_in_cell + idx_x)
