+++ tests/smoke_tests/parcel/arabas_and_shima_2017/test_conservation.py
@@ -47,7 +47,7 @@ def test_water_mass_conservation(settings_idx, mass_of_dry_air, scheme, coord):
     # Assert
     ql = liquid_water_mixing_ratio(simulation)
     qt = simulation.particulator.environment["qv"].to_ndarray() + ql
-    significant = 6 if scheme == "GPU" else 14  # TODO #540
+    significant = 6 if scheme == "GPU" else 12  # TODO #540
     np.testing.assert_approx_equal(qt, qt0, significant)
     if scheme != "SciPy":
         assert simulation.particulator.products["S_max"].get() >= output["S"][-1]
