+++ PySDM/dynamics/condensation.py
@@ -10,7 +10,7 @@ from ..physics import si
 
 DEFAULTS = namedtuple("_", ("rtol_x", "rtol_RH", "cond_range", "schedule"))(
     rtol_x=1e-6,
-    rtol_RH=1e-4,
+    rtol_RH=1e-6,
     cond_range=(1e-4 * si.second, 1 * si.second),
     schedule="dynamic",
 )
