### 2022-08-01 5b1893d
cleaning up pyvinecopulib code (as will remain only in -examples repo)
