### 2022-08-01 a90bfdf
make newer pylint happy. closes #854
