+++ tests/smoke_tests/niedermeier_et_al_2013/test_temperature_profile.py
@@ -5,16 +5,15 @@ from PySDM_examples.Niedermeier_et_al_2014 import Settings, Simulation
 
 from PySDM import Formulae
 from PySDM.physics import si
-from PySDM.physics.heterogeneous_ice_nucleation_rate import abifm
-
-abifm.m = 54.48
-abifm.c = -10.67
 
 
 @pytest.mark.parametrize("initial_temperature", (280 * si.K, 270 * si.K))
 def test_temperature_profile(initial_temperature, plot=False):
     # arrange
-    formulae = Formulae(heterogeneous_ice_nucleation_rate="ABIFM")
+    formulae = Formulae(
+        heterogeneous_ice_nucleation_rate="ABIFM",
+        constants={"ABIFM_M": 54.48, "ABIFM_C": -10.67},
+    )
     settings = Settings(
         initial_temperature=initial_temperature, timestep=10 * si.s, formulae=formulae
     )
