### 2022-08-02 4815d1e
Add breakup case for equal multiplicities
