+++ tests/unit_tests/dynamics/collisions/test_sdm_breakup.py
@@ -21,302 +21,302 @@ from PySDM.products.size_spectral import ParticleVolumeVersusRadiusLogarithmSpec
 
 
 class TestSDMBreakup:
-    @staticmethod
-    @pytest.mark.parametrize(
-        "dt",
-        [
-            1 * si.s,
-            10 * si.s,
-        ],
-    )
-    def test_nonadaptive_same_results_regardless_of_dt(dt, backend_class=CPU):
-        # Arrange
-        attributes = {
-            "n": np.asarray([1, 1]),
-            "volume": np.asarray([100 * si.um**3, 100 * si.um**3]),
-        }
-        breakup = Breakup(
-            collision_kernel=ConstantK(1 * si.cm**3 / si.s),
-            fragmentation_function=AlwaysN(4),
-            adaptive=False,
-        )
-        nsteps = 10
+    # @staticmethod
+    # @pytest.mark.parametrize(
+    #     "dt",
+    #     [
+    #         1 * si.s,
+    #         10 * si.s,
+    #     ],
+    # )
+    # def test_nonadaptive_same_results_regardless_of_dt(dt, backend_class=CPU):
+    #     # Arrange
+    #     attributes = {
+    #         "n": np.asarray([1, 1]),
+    #         "volume": np.asarray([100 * si.um**3, 100 * si.um**3]),
+    #     }
+    #     breakup = Breakup(
+    #         collision_kernel=ConstantK(1 * si.cm**3 / si.s),
+    #         fragmentation_function=AlwaysN(4),
+    #         adaptive=False,
+    #     )
+    #     nsteps = 10
 
-        n_sd = len(attributes["n"])
-        builder = Builder(n_sd, backend_class())
-        builder.set_environment(Box(dv=1 * si.cm**3, dt=dt))
-        builder.add_dynamic(breakup)
-        particulator = builder.build(attributes=attributes, products=())
+    #     n_sd = len(attributes["n"])
+    #     builder = Builder(n_sd, backend_class())
+    #     builder.set_environment(Box(dv=1 * si.cm**3, dt=dt))
+    #     builder.add_dynamic(breakup)
+    #     particulator = builder.build(attributes=attributes, products=())
 
-        # Act
-        particulator.run(nsteps)
+    #     # Act
+    #     particulator.run(nsteps)
 
-        # Assert
-        assert (particulator.attributes["n"].to_ndarray() > 0).all()
-        assert (particulator.attributes["n"].to_ndarray() != attributes["n"]).any()
-        assert np.sum(particulator.attributes["n"].to_ndarray()) >= np.sum(
-            attributes["n"]
-        )
-        assert (
-            particulator.attributes["n"].to_ndarray() == np.array([1024, 1024])
-        ).all()
+    #     # Assert
+    #     assert (particulator.attributes["n"].to_ndarray() > 0).all()
+    #     assert (particulator.attributes["n"].to_ndarray() != attributes["n"]).any()
+    #     assert np.sum(particulator.attributes["n"].to_ndarray()) >= np.sum(
+    #         attributes["n"]
+    #     )
+    #     assert (
+    #         particulator.attributes["n"].to_ndarray() == np.array([1024, 1024])
+    #     ).all()
 
-    @staticmethod
-    @pytest.mark.parametrize(
-        "params",
-        [
-            {
-                "gamma": 1.0,
-                "rand": 1.0,
-            },
-            {"gamma": 1.0, "rand": 0.1},
-            pytest.param(
-                {"gamma": 1.0, "rand": 0.0}, marks=pytest.mark.xfail(strict=True)
-            ),
-        ],
-    )
-    def test_single_collision_bounce(params, backend_class=CPU):
-        # Arrange
-        n_sd = 2
-        builder = Builder(n_sd, backend_class())
[…]
