### 2022-08-02 dd26c39
Unit test for xi_k = xi_j
