### 2022-08-03 bbc2a9f
do garbage collection for every test on 32 bit machines aiming at avoiding Numba memory errors
