### 2022-08-03 1cd8ebd
remove gc-collection hack - does not help... [skip ci]
