+++ tests/__init__.py
@@ -1,10 +1,1 @@
-import gc
-import platform
 
-import pytest
-
-
-@pytest.fixture(autouse=True)
-def ensure_gc():
-    if platform.architecture()[0] != "32bit":
-        gc.collect()
