### 2022-08-03 168d2e3
limit testing on Appveyor to unit tests
