+++ appveyor.yml
@@ -40,7 +40,7 @@ install:
 
 test_script:
   - ps: |
-      iex "$env:PYTHON -m pytest -We -p no:unraisableexception"
+      iex "$env:PYTHON -m pytest -We -p no:unraisableexception tests/unit_tests"
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
 
 #on_finish:
