### 2022-08-10 6e10d96
add SYSTEM_VERSION_COMPAT=0 for pip install (https://github.com/actions/setup-python/issues/469)
