### 2022-08-10 2c2fe20
bump macos-latest to macos-12
