### 2022-08-10 ebcfb7c
Reformulate breakup to perform maximum one round in a given backend call
