+++ PySDM/backends/impl_numba/methods/collisions_methods.py
@@ -1,6 +1,9 @@
 """
 CPU implementation of backend methods for particle collisions
 """
+from audioop import mul
+from xml.sax.xmlreader import AttributesImpl
+
 import numba
 import numpy as np
 
@@ -50,6 +53,85 @@ def coalesce(  # pylint: disable=too-many-arguments
 
 @numba.njit(**{**conf.JIT_FLAGS, **{"parallel": False}})
 def break_up(  # pylint: disable=too-many-arguments
+    i,
+    j,
+    k,
+    cid,
+    multiplicity,
+    gamma,
+    attributes,
+    n_fragment,
+    max_multiplicity,
+    min_volume,
+    breakup_rate,
+    breakup_rate_deficit,
+    warn_overflows,
+    volume,
+):  # pylint: disable=too-many-branches
+    # TODO: update the arguments for the calling function
+
+    # 1. find the max gamma that can be supported in 1 time step, add to rate,
+    #    and add remainder to the deficit (limits: max_xi, max_volume, min_volume)
+    overflow_flag = False
+    transfer_jk = 0
+    divisor_jk = 1
+    gamma_tmp = 0
+    gamma_deficit = gamma[i]
+    for m in range(int(gamma[i])):
+        transfer_jk_test = transfer_jk + divisor_jk
+        divisor_jk_test = divisor_jk * n_fragment[i]
+        new_n = multiplicity[j] - transfer_jk_test * multiplicity[k]
+        new_v = (volume[k] + transfer_jk_test * volume[j]) / divisor_jk_test
+        print("m, new_n, new_v = ", m, new_n, new_v)
+        # check for overflow of multiplicity
+        if (
+            transfer_jk_test > max_multiplicity
+            or multiplicity[k] * divisor_jk_test > max_multiplicity
+        ):
+            atomic_add(breakup_rate_deficit, cid, gamma_deficit * multiplicity[k])
+            overflow_flag = True
+            break
+        # check for new_n > 0, max volume, min volume
+        elif new_n < 0 or new_v > max(volume[j], volume[k]) or new_v < min_volume:
+            atomic_add(breakup_rate_deficit, cid, gamma_deficit * multiplicity[k])
+            break
+
+        # all tests passed
+        transfer_jk = transfer_jk_test
+        divisor_jk = divisor_jk_test
+        gamma_tmp = m + 1
+        gamma_deficit = gamma[i] - gamma_tmp
+    # 2. Compute the new multiplicities and particle sizes, with rounding
+    new_n = round(multiplicity[j] - transfer_jk * multiplicity[k])
+    for a in range(0, len(attributes)):
+        attributes[a, k] += transfer_jk * attributes[a, j]
+        attributes[a, k] /= divisor_jk
+    if new_n > 0:
+        nj = new_n
+        nk = multiplicity[k] * divisor_jk
+    else:
+        nj = divisor_jk * multiplicity[k] / 2
+        nk = nj
+        for a in range(0, len(attributes)):
+            attributes[a, j] = attributes[a, k]
+    # add up the product
+    atomic_add(breakup_rate, cid, gamma_tmp * multiplicity[k])
+    # perform rounding as necessary
+    multiplicity[j] = round(nj)
+    multiplicity[k] = round(nk)
+    factor_j = nj / multiplicity[j]
+    factor_k = nk / multiplicity[k]
+    for a in range(0, len(attributes)):
+        attributes[a, k] *= factor_k
+        attributes[a, j] *= factor_j
+
+    if overflow_flag:
+        if warn_overflows:
+            warn("overflow", __file__)
+
+
+@numba.njit(**{**conf.JIT_FLAGS, **{"parallel": False}})
+def break_up_old(  # pylint: disable=too-many-arguments
     i,
     j,
     k,
[…]
