+++ PySDM/particulator.py
@@ -143,6 +143,7 @@ class Particulator:  # pylint: disable=too-many-public-methods
         breakup_rate,
         breakup_rate_deficit,
         is_first_in_pair,
+        min_volume,
         warn_overflows,
     ):
         idx = self.attributes._ParticleAttributes__idx
@@ -166,7 +167,9 @@ class Particulator:  # pylint: disable=too-many-public-methods
                 breakup_rate=breakup_rate,
                 breakup_rate_deficit=breakup_rate_deficit,
                 is_first_in_pair=is_first_in_pair,
+                min_volume=min_volume,
                 warn_overflows=warn_overflows,
+                volume=self.attributes["volume"],
             )
         else:
             self.backend.collision_coalescence(
