### 2022-08-11 e63e73d
Remove debugging print statement
