### 2022-08-13 4955bb3
Add a switch and a test for the "while" version of breakup with no deficits
