+++ PySDM/particulator.py
@@ -145,6 +145,7 @@ class Particulator:  # pylint: disable=too-many-public-methods
         is_first_in_pair,
         min_volume,
         warn_overflows,
+        handle_all_breakups=False,
     ):
         idx = self.attributes._ParticleAttributes__idx
         healthy = self.attributes._ParticleAttributes__healthy_memory
@@ -170,6 +171,7 @@ class Particulator:  # pylint: disable=too-many-public-methods
                 min_volume=min_volume,
                 warn_overflows=warn_overflows,
                 volume=self.attributes["volume"],
+                handle_all_breakups=handle_all_breakups,
             )
         else:
             self.backend.collision_coalescence(
