+++ tests/unit_tests/dynamics/collisions/test_sdm_breakup.py
@@ -806,6 +806,109 @@ class TestSDMBreakup:
 
         assert (particulator.attributes["n"].to_ndarray() > 0).all()
 
+    @staticmethod
+    @pytest.mark.parametrize(
+        "params",
+        [
+            {
+                "gamma": [2.0],
+                "n_init": [1, 1],
+                "v_init": [1, 1],
+                "n_expected": [4, 4],
+                "v_expected": [0.25, 0.25],
+                "expected_deficit": [0.0],
+                "is_first_in_pair": [True, False],
+                "n_fragment": [4],
+            },
+            {
+                "gamma": [3.0],
+                "n_init": [9, 2],
+                "v_init": [1, 2],
+                "n_expected": [3, 17],
+                "v_expected": [5 / 9, 2 / 3],
+                "expected_deficit": [0.0],
+                "is_first_in_pair": [True, False],
+                "n_fragment": [3.0],
+            },
+        ],
+    )
+    @pytest.mark.parametrize("flag", ("n", "v", "conserve", "deficit"))
+    def test_while_loop_multi_breakup(params, flag, backend_class=CPU):
+        # Arrange
+        n_init = params["n_init"]
+        n_sd = len(n_init)
+        builder = Builder(n_sd, backend_class())
+        builder.set_environment(Box(dv=np.NaN, dt=np.NaN))
+        particulator = builder.build(
+            attributes={
+                "n": np.asarray(n_init),
+                "volume": np.asarray(params["v_init"]),
+            },
+            products=(),
+        )
+
+        n_pairs = n_sd // 2
+        rand = [1.0] * n_pairs
+        Eb = [1.0] * n_pairs
+        pairwise_zeros = particulator.PairwiseStorage.from_ndarray(
+            np.array([0.0] * n_pairs)
+        )
+        general_zeros = particulator.Storage.from_ndarray(np.array([0.0] * n_sd))
+
+        gamma = particulator.PairwiseStorage.from_ndarray(np.array(params["gamma"]))
+        rand = particulator.PairwiseStorage.from_ndarray(np.array(rand))
+        Eb = particulator.PairwiseStorage.from_ndarray(np.array(Eb))
+        breakup_rate = particulator.Storage.from_ndarray(np.array([0.0]))
+        breakup_rate_deficit = particulator.Storage.from_ndarray(np.array([0.0]))
+        n_fragment = particulator.PairwiseStorage.from_ndarray(
+            np.array(params["n_fragment"])
+        )
+        is_first_in_pair = particulator.PairIndicator(n_sd)
+        is_first_in_pair.indicator[:] = particulator.Storage.from_ndarray(
+            np.asarray(params["is_first_in_pair"])
+        )
+
+        # Act
+        particulator.collision_coalescence_breakup(
+            enable_breakup=True,
+            gamma=gamma,
+            rand=rand,
+            Ec=pairwise_zeros,
+            Eb=Eb,
+            n_fragment=n_fragment,
+            coalescence_rate=general_zeros,
+            breakup_rate=breakup_rate,
+            breakup_rate_deficit=breakup_rate_deficit,
+            is_first_in_pair=is_first_in_pair,
+            min_volume=1 * si.nm**3,
+            warn_overflows=True,
+            handle_all_breakups=True,
+        )
+
+        # Assert
+        {
+            "n": lambda: np.testing.assert_array_equal(
+                particulator.attributes["n"].to_ndarray(),
+                np.array(params["n_expected"]),
+            ),
+            "v": lambda: np.testing.assert_array_almost_equal(
+                particulator.attributes["volume"].to_ndarray(),
+                np.array(params["v_expected"]),
+                decimal=6,
+            ),
[…]
