### 2022-08-14 7b5f35f
enable Python 3.10 for CI
