+++ .github/workflows/tests+pypi.yml
@@ -19,7 +19,7 @@ jobs:
     strategy:
       matrix:
         platform: [ubuntu-latest, macos-12, windows-latest]
-        python-version: [3.7, 3.8, 3.9]
+        python-version: [3.7, 3.8, 3.9, 3.10]
     runs-on: ${{ matrix.platform }}
     steps:
       - uses: actions/checkout@v2
