### 2022-08-14 8e41998
bump SciPy requirement
