### 2022-08-14 e6c36a2
temporarily remove Python 3.10
