### 2022-08-15 3ff2245
down to SciPy 1.8.1
