### 2022-08-15 a507d2d
down to SciPy 1.7.3
