+++ tests/unit_tests/products/test_collision_rates.py
@@ -1,0 +1,149 @@
+# pylint: disable=missing-module-docstring,missing-class-docstring,missing-function-docstring
+from ast import Break
+
+import numpy as np
+import pytest
+
+import PySDM.physics.constants as const
+from PySDM import Builder
+from PySDM.backends import CPU
+from PySDM.backends.impl_common.pair_indicator import make_PairIndicator
+from PySDM.dynamics import Breakup, Coalescence, Collision
+from PySDM.dynamics.collisions.breakup_efficiencies import ConstEb
+from PySDM.dynamics.collisions.breakup_fragmentations import AlwaysN
+from PySDM.dynamics.collisions.coalescence_efficiencies import ConstEc
+from PySDM.dynamics.collisions.collision_kernels import ConstantK
+from PySDM.environments import Box
+from PySDM.physics import si
+from PySDM.products import (
+    BreakupRateDeficitPerGridbox,
+    BreakupRatePerGridbox,
+    CoalescenceRatePerGridbox,
+    CollisionRateDeficitPerGridbox,
+    CollisionRatePerGridbox,
+)
+
+
+class TestCollisionProducts:
+    @staticmethod
+    @pytest.mark.parametrize(
+        "params",
+        [
+            {
+                "gamma": 1.0,
+                "rand": 1.0,
+                "enable_breakup": False,
+                "cr": 2.0,
+                "crd": 0.0,
+                "cor": 2.0,
+                "br": 0.0,
+                "brd": 0.0,
+            },
+            # {
+            #     "gamma": 1.0,
+            #     "rand": 1.0,
+            #     "enable_breakup": True,
+            #     "enable_coalescence": True,
+            #     "Ec": 1.0,
+            # },
+            # {
+            #     "gamma": 1.0,
+            #     "rand": 1.0,
+            #     "enable_breakup": True,
+            #     "enable_coalescence": False,
+            #     "Ec": 0.0
+            # }
+        ],
+    )
+    def test_individual_dynamics_rates(params, backend_class=CPU):
+        # Arrange
+        n_init = [2, 5]
+        n_sd = len(n_init)
+        builder = Builder(n_sd, backend_class())
+        builder.set_environment(Box(dv=1 * si.m**3, dt=1 * si.s))
+        n_init = [1, 1]
+
+        # pairwise_zeros = particulator.PairwiseStorage.from_ndarray(np.array([0.0]))
+        # general_zeros = particulator.Storage.from_ndarray(np.array([0.0]))
+
+        # gamma = particulator.PairwiseStorage.from_ndarray(np.array([params["gamma"]]))
+        # rand = particulator.PairwiseStorage.from_ndarray(np.array([params["rand"]]))
+
+        if params["enable_breakup"]:
+            if params["enable_coalescence"]:
+                dynamic = Collision(
+                    collision_kernel=ConstantK(a=1 * si.cm**3 / si.s),
+                    coalescence_efficiency=ConstEc(Ec=params["Ec"]),
+                    breakup_efficiency=ConstEb(Eb=params["Eb"]),
+                    fragmentation_function=AlwaysN(n=params["nf"]),
+                )
+                products = (
+                    CollisionRatePerGridbox(name="cr"),
+                    CollisionRateDeficitPerGridbox(name="crd"),
+                    CoalescenceRatePerGridbox(name="cor"),
+                    BreakupRatePerGridbox(name="br"),
+                    BreakupRateDeficitPerGridbox(name="brd"),
+                )
+            else:
+                dynamic = Breakup(
+                    collision_kernel=ConstantK(a=1e5 * si.cm**3 / si.s),
+                    fragmentation_function=AlwaysN(n=params["nf"]),
+                )
+                products = (
+                    CollisionRatePerGridbox(name="cr"),
[…]
