### 2022-08-16 4857af5
more items in the DEFAULTS tuple to avoid value repetitions
