+++ PySDM/products/freezing/__init__.py
@@ -3,5 +3,13 @@ products pertinent to the `PySDM.dynamics.freezing.Freezing` dynamic
 """
 from .cooling_rate import CoolingRate
 from .freezable_specific_concentration import FreezableSpecificConcentration
+from .frozen_particle_concentration import (
+    FrozenParticleConcentration,
+    FrozenParticleSpecificConcentration,
+)
+from .ice_nuclei_concentration import (
+    IceNucleiConcentration,
+    SpecificIceNucleiConcentration,
+)
 from .ice_water_content import IceWaterContent, SpecificIceWaterContent
 from .total_unfrozen_immersed_surface_area import TotalUnfrozenImmersedSurfaceArea
