### 2022-08-18 ef9244e
Adding deficit rate tests for adaptivity settings
