+++ tests/unit_tests/products/test_collision_rates.py
@@ -25,125 +25,207 @@ from PySDM.products import (
 
 
 class TestCollisionProducts:
+    # @staticmethod
+    # @pytest.mark.parametrize(
+    #     "params",
+    #     [
+    #         { # coalescence only
+    #             "enable_breakup": False,
+    #             "cr": 4.0,
+    #             "crd": 6.0,
+    #             "cor": 4.0,
+    #         },
+    #         { # breakup and coalescence
+    #             "enable_breakup": True,
+    #             "enable_coalescence": True,
+    #             "Ec": 1.0,
+    #             "Eb": 1.0,
+    #             "nf": np.pi,
+    #             "cr": 4.0,
+    #             "crd": 6.0,
+    #             "cor": 4.0,
+    #             "br": 0.0,
+    #             "brd": 0.0,
+    #         },
+    #         { # breakup and coalescence
+    #             "enable_breakup": True,
+    #             "enable_coalescence": True,
+    #             "Ec": 0.0,
+    #             "Eb": 1.0,
+    #             "nf": 2.0,
+    #             "cr": 4.0,
+    #             "crd": 6.0,
+    #             "cor": 0.0,
+    #             "br": 2.0,
+    #             "brd": 2.0,
+    #         },
+    #         { # breakup only
+    #             "enable_breakup": True,
+    #             "enable_coalescence": False,
+    #             "nf": 2.0,
+    #             "cr": 4.0,
+    #             "crd": 6.0,
+    #             "br": 2.0,
+    #             "brd": 2.0,
+    #         },
+    #     ],
+    # )
+    # def test_individual_dynamics_rates_nonadaptive(params, backend_class=CPU):
+    #     # Arrange
+    #     n_init = [5, 2]
+    #     n_sd = len(n_init)
+    #     builder = Builder(n_sd, backend_class())
+    #     builder.set_environment(Box(dv=1 * si.m**3, dt=1 * si.s))
+
+    #     dynamic, productss = _get_dynamics_and_products(params, adaptive=False)
+    #     builder.add_dynamic(dynamic)
+
+    #     particulator = builder.build(
+    #         attributes={
+    #             "n": np.asarray(n_init),
+    #             "volume": np.asarray([100 * si.um**3] * n_sd),
+    #         },
+    #         products=productss,
+    #     )
+
+    #     # Act
+    #     particulator.run(1)
+
+    #     # Assert
+    #     assert particulator.products["cr"].get()[0] == params["cr"]
+    #     assert particulator.products["crd"].get()[0] == params["crd"]
+    #     if params["enable_breakup"]:
+    #         assert particulator.products["br"].get()[0] == params["br"]
+    #         assert particulator.products["brd"].get()[0] == params["brd"]
+    #         if params["enable_coalescence"]:
+    #             assert particulator.products["cor"].get()[0] == params["cor"]
+    #     else:
+    #         assert particulator.products["cor"].get()[0] == params["cor"]
+
+    # @staticmethod
+    # @pytest.mark.parametrize(
+    #     "params",
+    #     [
+    #         { # coalescence only
+    #             "enable_breakup": False,
+    #         },
+    #         { # breakup and coalescence
+    #             "enable_breakup": True,
+    #             "enable_coalescence": True,
+    #             "Ec": 1.0,
+    #             "Eb": 1.0,
+    #             "nf": np.pi,
+    #         },
+    #         { # breakup only
+    #             "enable_breakup": True,
+    #             "enable_coalescence": False,
+    #             "nf": np.pi,
+    #         },
+    #     ],
+    # )
+    # @pytest.mark.parametrize(
+    #     "n_init",
+    #     [
+    #         [5, 2],
+    #         [1, 2, 3, 4],
+    #         [3, 7] * 10
+    #     ],
[…]
