### 2022-08-18 7fe8440
Add test for sum of collision type rates
