+++ PySDM/backends/impl_numba/methods/collisions_methods.py
@@ -84,12 +84,10 @@ def break_up(  # pylint: disable=too-many-arguments,unused-argument
             transfer_jk_test > max_multiplicity
             or multiplicity[k] * divisor_jk_test > max_multiplicity
         ):
-            atomic_add(breakup_rate_deficit, cid, gamma_deficit * multiplicity[k])
             overflow_flag = True
             break
         # check for new_n > 0, max volume, min volume
         if new_n < 0 or new_v > max(volume[j], volume[k]) or new_v < min_volume:
-            atomic_add(breakup_rate_deficit, cid, gamma_deficit * multiplicity[k])
             break
 
         # all tests passed
@@ -112,6 +110,7 @@ def break_up(  # pylint: disable=too-many-arguments,unused-argument
             attributes[a, j] = attributes[a, k]
     # add up the product
     atomic_add(breakup_rate, cid, gamma_tmp * multiplicity[k])
+    atomic_add(breakup_rate_deficit, cid, gamma_deficit * multiplicity[k])
     # perform rounding as necessary
     multiplicity[j] = round(nj)
     multiplicity[k] = round(nk)
@@ -675,13 +674,6 @@ class CollisionsMethods(BackendMethods):
             atomic_add(collision_rate, cid, g * multiplicity[k])
             atomic_add(
                 collision_rate_deficit, cid, (int(gamma[i]) - g) * multiplicity[k]
-            )
-            print(
-                gamma[i],
-                g,
-                multiplicity[k],
-                collision_rate[cid],
-                collision_rate_deficit[cid],
             )
             gamma[i] = g
 
