+++ tests/unit_tests/products/test_collision_rates.py
@@ -25,136 +25,82 @@ from PySDM.products import (
 
 
 class TestCollisionProducts:
-    # @staticmethod
-    # @pytest.mark.parametrize(
-    #     "params",
-    #     [
-    #         { # coalescence only
-    #             "enable_breakup": False,
-    #             "cr": 4.0,
-    #             "crd": 6.0,
-    #             "cor": 4.0,
-    #         },
-    #         { # breakup and coalescence
-    #             "enable_breakup": True,
-    #             "enable_coalescence": True,
-    #             "Ec": 1.0,
-    #             "Eb": 1.0,
-    #             "nf": np.pi,
-    #             "cr": 4.0,
-    #             "crd": 6.0,
-    #             "cor": 4.0,
-    #             "br": 0.0,
-    #             "brd": 0.0,
-    #         },
-    #         { # breakup and coalescence
-    #             "enable_breakup": True,
-    #             "enable_coalescence": True,
-    #             "Ec": 0.0,
-    #             "Eb": 1.0,
-    #             "nf": 2.0,
-    #             "cr": 4.0,
-    #             "crd": 6.0,
-    #             "cor": 0.0,
-    #             "br": 2.0,
-    #             "brd": 2.0,
-    #         },
-    #         { # breakup only
-    #             "enable_breakup": True,
-    #             "enable_coalescence": False,
-    #             "nf": 2.0,
-    #             "cr": 4.0,
-    #             "crd": 6.0,
-    #             "br": 2.0,
-    #             "brd": 2.0,
-    #         },
-    #     ],
-    # )
-    # def test_individual_dynamics_rates_nonadaptive(params, backend_class=CPU):
-    #     # Arrange
-    #     n_init = [5, 2]
-    #     n_sd = len(n_init)
-    #     builder = Builder(n_sd, backend_class())
-    #     builder.set_environment(Box(dv=1 * si.m**3, dt=1 * si.s))
+    @staticmethod
+    @pytest.mark.parametrize(
+        "params",
+        [
+            {  # coalescence only
+                "enable_breakup": False,
+                "cr": 4.0,
+                "crd": 6.0,
+                "cor": 4.0,
+            },
+            {  # breakup and coalescence
+                "enable_breakup": True,
+                "enable_coalescence": True,
+                "Ec": 1.0,
+                "Eb": 1.0,
+                "nf": np.pi,
+                "cr": 4.0,
+                "crd": 6.0,
+                "cor": 4.0,
+                "br": 0.0,
+                "brd": 0.0,
+            },
+            {  # breakup and coalescence
+                "enable_breakup": True,
+                "enable_coalescence": True,
+                "Ec": 0.0,
+                "Eb": 1.0,
+                "nf": 2.0,
+                "cr": 4.0,
+                "crd": 6.0,
+                "cor": 0.0,
+                "br": 2.0,
+                "brd": 2.0,
+            },
+            {  # breakup only
+                "enable_breakup": True,
+                "enable_coalescence": False,
+                "nf": 2.0,
+                "cr": 4.0,
+                "crd": 6.0,
+                "br": 2.0,
+                "brd": 2.0,
+            },
+        ],
+    )
+    def test_individual_dynamics_rates_nonadaptive(params, backend_class=CPU):
+        # Arrange
+        n_init = [5, 2]
+        n_sd = len(n_init)
+        builder = Builder(n_sd, backend_class())
+        builder.set_environment(Box(dv=1 * si.m**3, dt=1 * si.s))
 
-    #     dynamic, productss = _get_dynamics_and_products(params, adaptive=False)
-    #     builder.add_dynamic(dynamic)
+        dynamic, productss = _get_dynamics_and_products(params, adaptive=False)
+        builder.add_dynamic(dynamic)
 
[…]
