### 2022-08-18 8e59dd8
Add back the coalescence only backend
