### 2022-08-18 01f63a2
Add unit test for multicell products to check indexing
