### 2022-08-18 59619fe
Fix smoke test for edjM
