### 2022-08-18 281fb0e
Fix smoke test edjM
