### 2022-08-19 f24c84d
add Fierce diagrams as a test for differences between full and linearised kappa-Koehler formulae
