+++ PySDM/backends/impl_numba/methods/collisions_methods.py
@@ -9,7 +9,9 @@ from PySDM.backends.impl_numba import conf
 from PySDM.backends.impl_numba.atomic_operations import atomic_add
 from PySDM.backends.impl_numba.storage import Storage
 from PySDM.backends.impl_numba.warnings import warn
-from PySDM.physics.constants import sqrt_pi, sqrt_two
+from PySDM.physics.constants import PI, PI_4_3, si, sqrt_pi, sqrt_two
+
+CM = si.cm
 
 
 @numba.njit(**{**conf.JIT_FLAGS, **{"parallel": False}})
@@ -218,6 +220,110 @@ def break_up_while(  # pylint: disable=too-many-arguments,unused-argument
         for a in range(0, len(attributes)):
             attributes[a, k] *= factor_k
             attributes[a, j] *= factor_j
+
+
+@numba.njit(**{**conf.JIT_FLAGS, **{"parallel": False}})
+def straub_Nr(  # pylint: disable=too-many-arguments,unused-argument
+    i,
+    CW,
+    gam,
+):  # pylint: disable=too-many-branches`
+    Nr = [0.0, 0.0, 0.0, 1.0, 1.0]
+    if gam[i] * CW[i] >= 7.0:
+        Nr[0] = 0.088 * (gam[i] * CW[i] - 7.0)
+    if CW[i] >= 21.0:
+        Nr[1] = 0.22 * (CW[i] - 21.0)
+        if CW[i] <= 46.0:
+            Nr[2] = 0.04 * (46.0 - CW[i])
+    else:
+        Nr[2] = 1.0
+
+    Nr[4] += Nr[0] + Nr[1] + Nr[2]
+
+    return Nr
+
+
+@numba.njit(**{**conf.JIT_FLAGS, **{"parallel": False}})
+def straub_p1(  # pylint: disable=too-many-arguments,unused-argument
+    i,
+    CW,
+    frag_size,
+    rand,
+):
+    E_D1 = 0.04 * CM
+    delD1 = 0.125 * CW[i] ** (1 / 2)
+    var_1 = delD1**2 / 12
+    sigma1 = np.sqrt(np.log(var_1 / E_D1**2 + 1))
+    mu1 = np.log(E_D1) - sigma1**2 / 2
+    X = rand[(i + 1) % len(CW)]
+
+    frag_size[i] = np.exp(
+        mu1
+        - sigma1 / sqrt_two / sqrt_pi / np.log(2) * np.log((1 / 2 + X) / (3 / 2 - X))
+    )
+    frag_size[i] = PI / 6 * frag_size[i] ** 3
+
+
+@numba.njit(**{**conf.JIT_FLAGS, **{"parallel": False}})
+def straub_p2(  # pylint: disable=too-many-arguments,unused-argument
+    i,
+    CW,
+    frag_size,
+    rand,
+):
+    mu2 = 0.095 * CM
+    delD2 = 0.22 * (CW[i] - 21.0)
+    sigma2 = delD2**2 / 12
+    X = rand[(i + 1) % len(CW)]
+
+    frag_size[i] = mu2 - sigma2 / sqrt_two / sqrt_pi / np.log(2) * np.log(
+        (1 / 2 + X) / (3 / 2 - X)
+    )
+    frag_size[i] = PI / 6 * frag_size[i] ** 3
+
+
+@numba.njit(**{**conf.JIT_FLAGS, **{"parallel": False}})
+def straub_p3(  # pylint: disable=too-many-arguments,unused-argument
+    i,
+    CW,
+    ds,
+    frag_size,
+    rand,
+):
+    mu3 = 0.9 * ds[i]
[…]
