+++ tests/smoke_tests/partmc/test_dry_wet_equilibration.py
@@ -1,0 +1,130 @@
+""" comparing kappa-Koehler wet radius equilibration in PySDM and PartMC
+  (based on PyPartMC-examples notebook by Zach D'Aquino) """
+# pylint: disable=missing-function-docstring
+import platform
+from collections import namedtuple
+
+import numpy as np
+import pytest
+from matplotlib import pyplot
+
+from PySDM import Builder
+from PySDM.backends import CPU
+from PySDM.environments import Box
+from PySDM.initialisation import equilibrate_wet_radii
+from PySDM.initialisation.spectra import Lognormal
+from PySDM.physics import si
+
+if platform.architecture()[0] == "64bit":
+    import PyPartMC
+
+
+def pysdm(dry_diam, temp, rel_humid, kpa):
+    r_dry = dry_diam / 2
+    builder = Builder(n_sd=0, backend=CPU())
+    environment = Box(dt=np.nan, dv=np.nan)
+    environment.register(builder)
+    environment["T"] = temp
+    environment["RH"] = rel_humid
+    kappa_times_dry_volume = kpa * (np.pi / 6) * dry_diam**3
+    return 2 * equilibrate_wet_radii(
+        r_dry=r_dry,
+        environment=environment,
+        kappa_times_dry_volume=kappa_times_dry_volume,
+    )
+
+
+def pypartmc(dry_diam, temp, rel_humid, kpa):
+    env_state = PyPartMC.EnvState(
+        {
+            "rel_humidity": rel_humid,
+            "latitude": 0.0,
+            "longitude": 0.0,
+            "altitude": 0.0,
+            "start_time": 0.0,
+            "start_day": 0,
+        }
+    )
+
+    env_state.set_temperature(temp)
+
+    aero_data = PyPartMC.AeroData(
+        (
+            {"H2O": [1000 * si.kg / si.m**3, 0, 18e-3 * si.kg / si.mol, 0]},
+            {"XXX": [np.nan * si.kg / si.m**3, 0, np.nan * si.kg / si.mol, kpa]},
+        )
+    )
+
+    dry_volumes = (np.pi / 6) * dry_diam**3
+    aero_particles = [
+        PyPartMC.AeroParticle(aero_data, np.array([0, 1]) * volume)
+        for volume in dry_volumes
+    ]
+
+    for aero_particle in aero_particles:
+        PyPartMC.condense_equilib_particle(env_state, aero_data, aero_particle)
+
+    wet_volumes = [np.sum(particle.volumes) for particle in aero_particles]
+    wet_diameters = ((6 / np.pi) * np.asarray(wet_volumes)) ** (1 / 3)
+
+    return wet_diameters
+
+
+@pytest.mark.skipif(
+    platform.architecture()[0] != "64bit", reason="binary package availability"
+)
+@pytest.mark.parametrize("kappa", (0.1, 1))
+@pytest.mark.parametrize("temperature", (300 * si.K,))
+@pytest.mark.parametrize("relative_humidity", (0.5, 0.75, 1.0))
+def test_dry_wet_equilibration(kappa, temperature, relative_humidity, plot=False):
+    # arrange
+    models = {"PySDM": pysdm, "PyPartMC": pypartmc}
+
+    Mode = namedtuple("Mode", ("norm_factor", "m_mode", "s_geom"))
+    modes = (
+        Mode(norm_factor=50000 / si.cm**3, m_mode=0.9 * si.um, s_geom=1.3),
+        Mode(norm_factor=80000 / si.cm**3, m_mode=5.8 * si.um, s_geom=2),
+    )
+
+    dry_diameters = np.logspace(-0.5, 1.5, 100) * si.um
[…]
