### 2022-08-23 caf4d90
fix errors with test and also fix error with szy-lang surface tension where sigma was undefined for f_org=0.0
