### 2022-08-23 b9e6407
fix errors with numba warning and plot xlim
