+++ .github/workflows/tests+pypi.yml
@@ -21,6 +21,7 @@ jobs:
         platform: [ubuntu-latest, macos-12, windows-latest]
         python-version: ["3.7", "3.8", "3.9"]
     runs-on: ${{ matrix.platform }}
+    timeout-minutes: 90
     steps:
       - uses: actions/checkout@v2
 
@@ -58,6 +59,7 @@ jobs:
 
   dist:
     runs-on: ubuntu-latest
+    timeout-minutes: 90
     needs: [build]
     steps:
       - uses: actions/checkout@v2
