### 2022-08-24 21edb78
add GITHUB_TOKEN to pytest run (and clean up THREADING_LAYER setup)
