+++ .github/workflows/tests+pypi.yml
@@ -44,12 +44,11 @@ jobs:
           pip install -e .
 
       # https://github.com/numba/numba/issues/6350#issuecomment-728174860
-      - if: matrix.platform == 'ubuntu-latest'
-        env: 
-          NUMBA_THREADING_LAYER: omp
-        run: pytest -p no:unraisableexception -We
+      - if: startsWith(matrix.platform, 'ubuntu-')
+        run: echo NUMBA_THREADING_LAYER=omp >> $GITHUB_ENV
 
-      - if: matrix.platform != 'ubuntu-latest'
+      - env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         run: pytest -p no:unraisableexception -We
 
 #       - if: ${{ failure() }}
