### 2022-08-24 fede658
use secrets.GITHUB_TOKEN for ghapi in devops tests
