### 2022-08-24 4eb3a97
use -We flag for pytest in codecov run
