### 2022-08-24 be9c9b5
WIP on Straub 2010
