+++ PySDM/backends/impl_numba/methods/collisions_methods.py
@@ -225,22 +225,25 @@ def break_up_while(  # pylint: disable=too-many-arguments,unused-argument
 @numba.njit(**{**conf.JIT_FLAGS, **{"parallel": False}})
 def straub_Nr(  # pylint: disable=too-many-arguments,unused-argument
     i,
+    Nr1,
+    Nr2,
+    Nr3,
+    Nr4,
+    Nrt,
     CW,
     gam,
 ):  # pylint: disable=too-many-branches`
-    Nr = [0.0, 0.0, 0.0, 1.0, 1.0]
+    # Nr = [0.0, 0.0, 0.0, 1.0, 1.0]
     if gam[i] * CW[i] >= 7.0:
-        Nr[0] = 0.088 * (gam[i] * CW[i] - 7.0)
+        Nr1[i] = 0.088 * (gam[i] * CW[i] - 7.0)
     if CW[i] >= 21.0:
-        Nr[1] = 0.22 * (CW[i] - 21.0)
+        Nr2[i] = 0.22 * (CW[i] - 21.0)
         if CW[i] <= 46.0:
-            Nr[2] = 0.04 * (46.0 - CW[i])
+            Nr3[i] = 0.04 * (46.0 - CW[i])
     else:
-        Nr[2] = 1.0
-
-    Nr[4] += Nr[0] + Nr[1] + Nr[2]
-
-    return Nr
+        Nr3[i] = 1.0
+    Nr4[i] = 1.0
+    Nrt[i] = Nr1[i] + Nr2[i] + Nr3[i] + Nr4[i]
 
 
 @numba.njit(**{**conf.JIT_FLAGS, **{"parallel": False}})
@@ -303,7 +306,7 @@ def straub_p3(  # pylint: disable=too-many-arguments,unused-argument
 
 @numba.njit(**{**conf.JIT_FLAGS, **{"parallel": False}})
 def straub_p4(  # pylint: disable=too-many-arguments,unused-argument
-    i, CW, ds, v_max, frag_size, Nr
+    i, CW, ds, v_max, frag_size, Nr1, Nr2, Nr3
 ):
     E_D1 = 0.04 * CM
     delD1 = 0.125 * CW[i] ** (1 / 2)
@@ -318,9 +321,9 @@ def straub_p4(  # pylint: disable=too-many-arguments,unused-argument
     sigma3 = delD3**2 / 12
     dl = 2 * (v_max[i] / PI_4_3) ** (1 / 3)
 
-    M31 = Nr[0] * np.exp(3 * mu1 + 9 * sigma1**2 / 2)
-    M32 = Nr[1] * (mu2**3 + 3 * mu2 * sigma2**2)
-    M33 = Nr[2] * (mu3**3 + 3 * mu3 * sigma3**2)
+    M31 = Nr1[i] * np.exp(3 * mu1 + 9 * sigma1**2 / 2)
+    M32 = Nr2[i] * (mu2**3 + 3 * mu2 * sigma2**2)
+    M33 = Nr3[i] * (mu3**3 + 3 * mu3 * sigma3**2)
 
     M34 = v_max[i] / PI_4_3 * 8 + ds[i] ** 3 - M31 - M32 - M33
     frag_size[i] = PI / 6 * M34
@@ -600,8 +603,10 @@ class CollisionsMethods(BackendMethods):
             n_fragment[i] = x_plus_y[i] / frag_size[i]
             if frag_size[i] > v_max[i]:
                 n_fragment[i] = 1
+                frag_size[i] = x_plus_y[i]
             elif frag_size[i] < vmin:
                 n_fragment[i] = 1
+                frag_size[i] = x_plus_y[i]
 
             if nfmax is not None:
[…]
