### 2022-08-24 82bed81
remove parallel computation from test_zero_forg. hopefully this helps with the timeout issues on mac and windows.
