+++ tests/smoke_tests/lowe_et_al_2019/test_zero_forg.py
@@ -1,7 +1,5 @@
 # pylint: disable=missing-module-docstring,missing-class-docstring,missing-function-docstring
 import numpy as np
-import pytest
-from joblib import Parallel, delayed, parallel_backend
 from matplotlib import pyplot
 from PySDM_examples.Lowe_et_al_2019 import Settings, Simulation
 from PySDM_examples.Lowe_et_al_2019.aerosol import AerosolBoreal, AerosolMarine
@@ -10,17 +8,7 @@ from PySDM.initialisation.sampling import spectral_sampling as spec_sampling
 from PySDM.physics import si
 
 
-@pytest.mark.parametrize()
-def compute(key, settings):
-    simulation = Simulation(settings)
-    output = simulation.run()
-    output["updraft"] = settings.w
-    output["org_fraction"] = settings.aerosol.modes[0]["f_org"]
-    output["color"] = settings.aerosol.color
-    return key, output
-
-
-def test_zero_forg(plot=True):
+def test_zero_forg(plot=False):
     nRes = 3
     updraft_list = np.geomspace(0.1, 10, nRes)
 
@@ -38,33 +26,27 @@ def test_zero_forg(plot=True):
         "BDF": False,
     }
 
-    with parallel_backend("loky", n_jobs=1):
-        output = dict(
-            Parallel(verbose=10)(
-                delayed(compute)(
-                    subplot + f"_w{w:.2f}_" + model,
-                    Settings(
-                        dz=1 * si.m,
-                        n_sd_per_mode=20,
-                        model=model,
-                        aerosol={
-                            "a": AerosolMarine(Forg=0, Acc_N2=Acc["a"]),
-                            "b": AerosolMarine(Forg=0, Acc_N2=Acc["b"]),
-                            "c": AerosolBoreal(Forg=0, Acc_N2=Acc["c"]),
-                            "d": AerosolBoreal(Forg=0, Acc_N2=Acc["d"]),
-                        }[subplot],
-                        w=w * si.m / si.s,
-                        spectral_sampling=spec_sampling.ConstantMultiplicity,
-                        **consts,
-                    ),
-                )
-                for w in updraft_list
-                for subplot in subplot_list
-                for model in models
-            )
-        )
-
     cdnc_compare = np.zeros((len(models), len(subplot_list), len(updraft_list)))
+    for i, w in enumerate(updraft_list):
+        for k, subplot in enumerate(subplot_list):
+            for m, model in enumerate(models):
+                settings = Settings(
+                    dz=1 * si.m,
+                    n_sd_per_mode=20,
+                    model=model,
+                    aerosol={
+                        "a": AerosolMarine(Forg=0, Acc_N2=Acc["a"]),
+                        "b": AerosolMarine(Forg=0, Acc_N2=Acc["b"]),
+                        "c": AerosolBoreal(Forg=0, Acc_N2=Acc["c"]),
+                        "d": AerosolBoreal(Forg=0, Acc_N2=Acc["d"]),
+                    }[subplot],
+                    w=w * si.m / si.s,
+                    spectral_sampling=spec_sampling.ConstantMultiplicity,
+                    **consts,
+                )
+                simulation = Simulation(settings)
+                output = simulation.run()
+                cdnc_compare[m, k, i] = np.array(output["n_c_cm3"])[-1]
 
     mrkr = ["o", "s", "*", "v", "^", "D", "h", "x", "+", "8", "p", "<", ">", "d", "H"]
     _, axes = pyplot.subplots(
@@ -84,20 +66,20 @@ def test_zero_forg(plot=True):
 
         for m, model in enumerate(models):
             for i, w in enumerate(updraft_list):
-                key = subplot + f"_w{w:.2f}_"
-                CDNC = np.array(output[key + model]["n_c_cm3"])
[…]
