### 2022-08-25 f4eddd0
conditional calculation of sgm in ruehl comp film if f_org > 0, if no organics, then sigma=sgm_w
