### 2022-08-26 e1afb03
averaged terminal velocity product implemented; unit tests added; tests pass
