### 2022-08-26 dcfec1f
add -We to pytest run in codecov workflow. closes #897
