### 2022-08-29 5356b8c
increase minimum bracket interval. maybe this will avoid the overflow in exp for ruehl surface tension.
