+++ PySDM/physics/surface_tension/compressed_film_ruehl.py
@@ -13,9 +13,9 @@ from PySDM.physics.trivia import Trivia
 # pylint: disable=too-many-arguments
 @numba.njit(**{**jit_flags, "parallel": False})
 def minfun(f_surf, Cb_iso, RUEHL_C0, RUEHL_A0, A_iso, c):
-    return Cb_iso * (1 - f_surf) / RUEHL_C0 - np.exp(
-        c * (RUEHL_A0**2 - (A_iso / f_surf) ** 2)
-    )
+    lhs = np.log(Cb_iso * (1 - f_surf) / RUEHL_C0)
+    rhs = c * (RUEHL_A0**2 - (A_iso / f_surf) ** 2)
+    return lhs - rhs
 
 
 within_tolerance = numba.njit(
@@ -45,10 +45,14 @@ class CompressedFilmRuehl:
 
     @staticmethod
     def sigma(const, T, v_wet, v_dry, f_org):
+        assert 0 <= f_org <= 1
+
         # wet radius (m)
         r_wet = ((3 * v_wet) / (4 * const.PI)) ** (1 / 3)
 
-        if f_org > 0:
+        if f_org == 0:
+            sgm = const.sgm_w
+        else:
             # C_bulk is the concentration of the organic in the bulk phase
             # Cb_iso = C_bulk / (1-f_surf)
             Cb_iso = (f_org * v_dry / const.RUEHL_nu_org) / (v_wet / const.nu_w)
@@ -65,7 +69,7 @@ class CompressedFilmRuehl:
             args = (Cb_iso, const.RUEHL_C0, const.RUEHL_A0, A_iso, c)
             rtol = 1e-6
             max_iters = 1e2
-            bracket = (1e-6, 1)
+            bracket = (rtol, 1 - rtol)
             f_surf, iters = toms748_solve(
                 minfun,
                 args,
@@ -80,8 +84,6 @@ class CompressedFilmRuehl:
 
             # calculate surface tension
             sgm = const.sgm_w - (const.RUEHL_A0 - A_iso / f_surf) * const.RUEHL_m_sigma
-        else:
-            sgm = const.sgm_w
 
         # surface tension bounded between sgm_min and sgm_w
         sgm = np.minimum(np.maximum(sgm, const.RUEHL_sgm_min), const.sgm_w)
