+++ PySDM/physics/surface_tension/szyszkowski_langmuir.py
@@ -25,35 +25,40 @@ class SzyszkowskiLangmuir:
 
     @staticmethod
     def sigma(const, T, v_wet, v_dry, f_org):
-        r_wet = ((3 * v_wet) / (4 * np.pi)) ** (1 / 3)  # m - wet radius
+        assert 0 <= f_org <= 1
 
-        # C_bulk is the concentration of the organic in the bulk phase
-        # Cb_iso = C_bulk / (1-f_surf)
-        Cb_iso = (f_org * v_dry / const.RUEHL_nu_org) / (v_wet / const.nu_w)
+        # wet radius (m)
+        r_wet = ((3 * v_wet) / (4 * np.pi)) ** (1 / 3)
 
-        # A is the area that one molecule of organic occupies at the droplet surface
-        # A_iso = A*f_surf (m^2)
-        A_iso = (4 * np.pi * r_wet**2) / (
-            f_org * v_dry * const.N_A / const.RUEHL_nu_org
-        )
-
-        # fraction of organic at surface
-        # quadratic formula, solve equation of state analytically
-        a = -const.RUEHL_A0 / A_iso
-        b = (
-            const.RUEHL_A0 / A_iso
-            + (const.RUEHL_A0 / A_iso) * (const.RUEHL_C0 / Cb_iso)
-            + 1
-        )
-        c = -1
-        if f_org > 0:
-            f_surf = (-b + np.sqrt(b**2 - 4 * a * c)) / (2 * a)
+        if f_org == 0:
+            sgm = const.sgm_w
         else:
-            f_surf = 0
+            # C_bulk is the concentration of the organic in the bulk phase
+            # Cb_iso = C_bulk / (1-f_surf)
+            Cb_iso = (f_org * v_dry / const.RUEHL_nu_org) / (v_wet / const.nu_w)
 
-        # calculate surface tension
-        sgm = const.sgm_w - ((const.R_str * T) / (const.RUEHL_A0 * const.N_A)) * np.log(
-            1 + Cb_iso * (1 - f_surf) / const.RUEHL_C0
-        )
+            # A is the area that one molecule of organic occupies at the droplet surface
+            # A_iso = A*f_surf (m^2)
+            A_iso = (4 * np.pi * r_wet**2) / (
+                f_org * v_dry * const.N_A / const.RUEHL_nu_org
+            )
+
+            # fraction of organic at surface
+            # quadratic formula, solve equation of state analytically
+            a = -const.RUEHL_A0 / A_iso
+            b = (
+                const.RUEHL_A0 / A_iso
+                + (const.RUEHL_A0 / A_iso) * (const.RUEHL_C0 / Cb_iso)
+                + 1
+            )
+            c = -1
+            f_surf = (-b + np.sqrt(b**2 - 4 * a * c)) / (2 * a)
+
+            # calculate surface tension
+            sgm = const.sgm_w - (
+                (const.R_str * T) / (const.RUEHL_A0 * const.N_A)
+            ) * np.log(1 + Cb_iso * (1 - f_surf) / const.RUEHL_C0)
+
+        # surface tension bounded between sgm_min and sgm_w
         sgm = np.minimum(np.maximum(sgm, const.RUEHL_sgm_min), const.sgm_w)
         return sgm
