### 2022-08-29 6ce050a
fixes to make shipway and hill smoke tests pass.
