### 2022-08-29 17e13cd
[skip ci] explicit treatment of edge cases for ruehl
