+++ PySDM/physics/surface_tension/szyszkowski_langmuir.py
@@ -25,8 +25,6 @@ class SzyszkowskiLangmuir:
 
     @staticmethod
     def sigma(const, T, v_wet, v_dry, f_org):
-        assert 0 <= f_org <= 1
-
         # wet radius (m)
         r_wet = ((3 * v_wet) / (4 * np.pi)) ** (1 / 3)
 
