### 2022-08-31 1e99963
Reformulate breakup efficiency as multiplicative
