### 2022-08-31 e3c6863
Rescale random numbers in straub 2010 for second step
