### 2022-08-31 55ef301
Reformulate breakup in terms of fragment size only
