+++ PySDM/dynamics/collisions/breakup_fragmentations/exponential.py
@@ -11,7 +11,6 @@ class ExponFrag:
         self.vmin = vmin
         self.nfmax = nfmax
         self.max_size = None
-        self.frag_size = None
         self.sum_of_volumes = None
 
     def register(self, builder):
@@ -19,22 +18,19 @@ class ExponFrag:
         self.max_size = self.particulator.PairwiseStorage.empty(
             self.particulator.n_sd // 2, dtype=float
         )
-        self.frag_size = self.particulator.PairwiseStorage.empty(
-            self.particulator.n_sd // 2, dtype=float
-        )
         self.sum_of_volumes = self.particulator.PairwiseStorage.empty(
             self.particulator.n_sd // 2, dtype=float
         )
 
-    def __call__(self, output, u01, is_first_in_pair):
+    def __call__(self, nf, frag_size, u01, is_first_in_pair):
         self.max_size.max(self.particulator.attributes["volume"], is_first_in_pair)
         self.sum_of_volumes.sum(
             self.particulator.attributes["volume"], is_first_in_pair
         )
         self.particulator.backend.exp_fragmentation(
-            n_fragment=output,
+            n_fragment=nf,
             scale=self.scale,
-            frag_size=self.frag_size,
+            frag_size=frag_size,
             v_max=self.max_size,
             x_plus_y=self.sum_of_volumes,
             rand=u01,
