+++ PySDM/backends/impl_numba/methods/collisions_methods.py
@@ -2,6 +2,7 @@
 CPU implementation of backend methods for particle collisions
 """
 from audioop import mul
+from xml.dom.expatbuilder import FragmentBuilder
 
 import numba
 import numpy as np
@@ -64,7 +65,6 @@ def break_up(  # pylint: disable=too-many-arguments,unused-argument
     n_fragment,
     fragment_size,
     max_multiplicity,
-    min_volume,
     breakup_rate,
     breakup_rate_deficit,
     warn_overflows,
@@ -76,27 +76,29 @@ def break_up(  # pylint: disable=too-many-arguments,unused-argument
     #    and add remainder to the deficit (limits: max_multiplicity, min_volume,
     #    volume doesn't exceed initial particle volumes)
     overflow_flag = False
-    take_from_j_test = 1 - n_fragment[i]
+    take_from_j_test = multiplicity[k]
     new_mult_k_test = 0
+    new_mult_k = multiplicity[k]
+    take_from_j = 0
     gamma_tmp = 0
     gamma_deficit = gamma[i]
     for m in range(int(gamma[i])):
+        take_from_j_test = new_mult_k_test + take_from_j_test
         new_mult_k_test = (
-            new_mult_k_test * (volume[j] / fragment_size[i]) + n_fragment[i]
+            new_mult_k_test * (volume[j] / fragment_size[i])
+            + n_fragment[i] * multiplicity[k]
         )
-        take_from_j_test = take_from_j_test + new_mult_k_test
-
         # check for overflow of multiplicity
         if new_mult_k_test > max_multiplicity:
             overflow_flag = True
             break
-        # check for new_n > 0, max volume, min volume
+        # check for new_n > 0
         if take_from_j_test > multiplicity[j]:
             break
 
         # all tests passed
         take_from_j = take_from_j_test
-        new_mult_k = new_mult_k
+        new_mult_k = new_mult_k_test
         gamma_tmp = m + 1
         gamma_deficit = gamma[i] - gamma_tmp
     # 2. Compute the new multiplicities and particle sizes, with rounding
@@ -109,9 +111,13 @@ def break_up(  # pylint: disable=too-many-arguments,unused-argument
         nk = new_mult_k
     else:
         nj = new_mult_k / 2
-        nk = nj
-        for a in range(0, len(attributes)):
-            attributes[a, j] = attributes[a, k]
+        if round(nj) == 0:
+            atomic_add(breakup_rate_deficit, cid, gamma[i] * multiplicity[k])
+            return
+        else:
+            nk = nj
+            for a in range(0, len(attributes)):
+                attributes[a, j] = attributes[a, k]
     # add up the product
     atomic_add(breakup_rate, cid, gamma_tmp * multiplicity[k])
     atomic_add(breakup_rate_deficit, cid, gamma_deficit * multiplicity[k])
@@ -139,8 +145,8 @@ def break_up_while(  # pylint: disable=too-many-arguments,unused-argument
     gamma,
     attributes,
     n_fragment,
+    fragment_size,
     max_multiplicity,
-    min_volume,
     breakup_rate,
     breakup_rate_deficit,
     warn_overflows,
@@ -151,69 +157,63 @@ def break_up_while(  # pylint: disable=too-many-arguments,unused-argument
     overflow_flag = False
     while gamma_deficit > 0:
         if multiplicity[k] == multiplicity[j]:
-            gamma_tmp = gamma_deficit
-            tmp2 = (n_fragment[i] / 2) ** gamma_tmp
-            new_n = multiplicity[k] * tmp2
-            if new_n > max_multiplicity:
+            take_from_j = multiplicity[j]
+            new_mult_k = (volume[j] + volume[k]) / fragment_size[i] * multiplicity[k]
[…]
