+++ tests/unit_tests/dynamics/collisions/test_sdm_breakup.py
@@ -1,4 +1,6 @@
 # pylint: disable=missing-module-docstring,missing-class-docstring,missing-function-docstring
+from xml.dom.expatbuilder import FragmentBuilder
+
 import numpy as np
 import pytest
 
@@ -95,6 +97,7 @@ class TestSDMBreakup:
         gamma = particulator.PairwiseStorage.from_ndarray(np.array([params["gamma"]]))
         rand = particulator.PairwiseStorage.from_ndarray(np.array([params["rand"]]))
         n_fragment = particulator.PairwiseStorage.from_ndarray(np.array([4]))
+        fragment_size = particulator.PairwiseStorage.from_ndarray(np.array([-1]))
         is_first_in_pair = make_PairIndicator(backend_class)(n_sd)
         min_volume = 1 * si.nm**3
 
@@ -106,6 +109,7 @@ class TestSDMBreakup:
             Ec=pairwise_zeros,
             Eb=pairwise_zeros,
             n_fragment=n_fragment,
+            fragment_size=fragment_size,
             coalescence_rate=general_zeros,
             breakup_rate=general_zeros,
             breakup_rate_deficit=general_zeros,
@@ -182,6 +186,7 @@ class TestSDMBreakup:
         )
         breakup_rate = particulator.Storage.from_ndarray(np.array([0.0]))
         n_fragment = particulator.PairwiseStorage.from_ndarray(np.array([4] * n_pairs))
+        frag_size = particulator.PairwiseStorage.from_ndarray(np.array([-1] * n_pairs))
         is_first_in_pair = particulator.PairIndicator(n_sd)
         is_first_in_pair.indicator[:] = particulator.Storage.from_ndarray(
             np.asarray(params["is_first_in_pair"])
@@ -196,6 +201,7 @@ class TestSDMBreakup:
             Ec=pairwise_zeros,
             Eb=Eb,
             n_fragment=n_fragment,
+            fragment_size=frag_size,
             coalescence_rate=general_zeros,
             breakup_rate=breakup_rate,
             breakup_rate_deficit=general_zeros,
@@ -223,16 +229,18 @@ class TestSDMBreakup:
                 "expected_deficit": [0.0],
                 "is_first_in_pair": [True, False],
                 "n_fragment": [4],
+                "frag_size": [0.5],
             },
             {
                 "gamma": [2.0],
                 "n_init": [20, 4],
                 "v_init": [1, 2],
-                "n_expected": [4, 36],
-                "v_expected": [1, 2 / 3],
+                "n_expected": [4, 24],
+                "v_expected": [1, 1],
                 "expected_deficit": [0.0],
                 "is_first_in_pair": [True, False],
                 "n_fragment": [3],
+                "frag_size": [1.0],
             },
             {
                 "gamma": [2.0],
@@ -243,6 +251,7 @@ class TestSDMBreakup:
                 "expected_deficit": [1.0],
                 "is_first_in_pair": [True, False],
                 "n_fragment": [4],
+                "frag_size": [0.5],
             },
             {
                 "gamma": [2.0],
@@ -253,6 +262,7 @@ class TestSDMBreakup:
                 "expected_deficit": [1.0],
                 "is_first_in_pair": [True, False],
                 "n_fragment": [4],
+                "frag_size": [0.5],
             },
         ],
     )
@@ -287,6 +297,9 @@ class TestSDMBreakup:
         n_fragment = particulator.PairwiseStorage.from_ndarray(
             np.array(params["n_fragment"])
         )
+        frag_size = particulator.PairwiseStorage.from_ndarray(
+            np.asarray(params["frag_size"])
+        )
         is_first_in_pair = particulator.PairIndicator(n_sd)
[…]
