+++ PySDM/dynamics/collisions/breakup_fragmentations/feingold1988.py
@@ -12,7 +12,6 @@ class Feingold1988Frag:
         self.vmin = vmin
         self.nfmax = nfmax
         self.max_size = None
-        self.frag_size = None
         self.sum_of_volumes = None
 
     def register(self, builder):
@@ -21,22 +20,19 @@ class Feingold1988Frag:
         self.max_size = self.particulator.PairwiseStorage.empty(
             self.particulator.n_sd // 2, dtype=float
         )
-        self.frag_size = self.particulator.PairwiseStorage.empty(
-            self.particulator.n_sd // 2, dtype=float
-        )
         self.sum_of_volumes = self.particulator.PairwiseStorage.empty(
             self.particulator.n_sd // 2, dtype=float
         )
 
-    def __call__(self, output, u01, is_first_in_pair):
+    def __call__(self, nf, frag_size, u01, is_first_in_pair):
         self.max_size.max(self.particulator.attributes["volume"], is_first_in_pair)
         self.sum_of_volumes.sum(
             self.particulator.attributes["volume"], is_first_in_pair
         )
         self.particulator.backend.feingold1988_fragmentation(
-            n_fragment=output,
+            n_fragment=nf,
             scale=self.scale,
-            frag_size=self.frag_size,
+            frag_size=frag_size,
             v_max=self.max_size,
             x_plus_y=self.sum_of_volumes,
             rand=u01,

[…]
