+++ PySDM/particulator.py
@@ -144,7 +144,6 @@ class Particulator:  # pylint: disable=too-many-public-methods
         breakup_rate,
         breakup_rate_deficit,
         is_first_in_pair,
-        min_volume,
         warn_overflows,
         handle_all_breakups=False,
     ):
@@ -170,7 +169,6 @@ class Particulator:  # pylint: disable=too-many-public-methods
                 breakup_rate=breakup_rate,
                 breakup_rate_deficit=breakup_rate_deficit,
                 is_first_in_pair=is_first_in_pair,
-                min_volume=min_volume,
                 warn_overflows=warn_overflows,
                 volume=self.attributes["volume"],
                 handle_all_breakups=handle_all_breakups,
