+++ PySDM/dynamics/collisions/collision.py
@@ -22,9 +22,8 @@ from PySDM.dynamics.impl.random_generator_optimizer_nopair import (
 )
 from PySDM.physics import si
 
-DEFAULTS = namedtuple("_", ("dt_coal_range", "min_volume", "adaptive", "substeps"))(
+DEFAULTS = namedtuple("_", ("dt_coal_range", "adaptive", "substeps"))(
     dt_coal_range=(0.1 * si.second, 100.0 * si.second),
-    min_volume=0.0,
     adaptive=True,
     substeps=1,
 )
@@ -44,7 +43,6 @@ class Collision:
         adaptive: bool = DEFAULTS.adaptive,
         dt_coal_range=DEFAULTS.dt_coal_range,
         enable_breakup: bool = True,
-        min_volume=DEFAULTS.min_volume,
         warn_overflows: bool = True,
         handle_all_breakups: bool = False,
     ):
@@ -56,7 +54,6 @@ class Collision:
         self.enable_breakup = enable_breakup
         self.warn_overflows = warn_overflows
         self.handle_all_breakups = handle_all_breakups
-        self.min_volume = min_volume
 
         self.collision_kernel = collision_kernel
         self.compute_coalescence_efficiency = coalescence_efficiency
@@ -216,7 +213,6 @@ class Collision:
             breakup_rate=self.breakup_rate,
             breakup_rate_deficit=self.breakup_rate_deficit,
             is_first_in_pair=self.is_first_in_pair,
-            min_volume=self.min_volume,
             warn_overflows=self.warn_overflows,
             handle_all_breakups=self.handle_all_breakups,
         )
@@ -319,7 +315,6 @@ class Breakup(Collision):
         substeps: int = DEFAULTS.substeps,
         adaptive: bool = DEFAULTS.adaptive,
         dt_coal_range=DEFAULTS.dt_coal_range,
-        min_volume=DEFAULTS.min_volume,
     ):
         coalescence_efficiency = ConstEc(Ec=0.0)
         breakup_efficiency = ConstEb(Eb=1.0)
@@ -333,5 +328,4 @@ class Breakup(Collision):
             substeps=substeps,
             adaptive=adaptive,
             dt_coal_range=dt_coal_range,
-            min_volume=min_volume,
         )
