### 2022-09-02 06770b4
EffectiveRadius refactor to avoid divide-by-zero warnings when run without Numba
