### 2022-09-02 82f5c2e
add `git diff` at the end of pre-commit hooks so that GA logs will show what was changed
