### 2022-09-02 d8108a8
fix EffectiveRadius init/register
