### 2022-09-06 6da668d
modify surface tension SL and Ruehl to work for scalars or arrays. remove loops from tests to ensure this works
