+++ tests/smoke_tests/lowe_et_al_2019/test_surface_tension_models.py
@@ -166,11 +166,9 @@ class TestFig1:
         )
 
         # act
-        sigma = np.zeros(len(V_WET))
-        for i, vw in enumerate(V_WET):
-            sigma[i] = formulae.surface_tension.sigma.py_func(
-                TEMPERATURE, vw, V_DRY, aer.modes[0]["f_org"]
-            )
+        sigma = formulae.surface_tension.sigma.py_func(
+            TEMPERATURE, V_WET, V_DRY, aer.modes[0]["f_org"]
+        )
 
         # assert
         test = np.array(
