+++ PySDM/physics/surface_tension/szyszkowski_langmuir.py
@@ -25,37 +25,48 @@ class SzyszkowskiLangmuir:
 
     @staticmethod
     def sigma(const, T, v_wet, v_dry, f_org):
+        # find where f_org=0
+        zero = (
+            np.equal(f_org, np.zeros_like(f_org))
+            * np.equal(v_wet, v_wet)
+            * np.equal(v_dry, v_dry)
+        )
+
+        # reset to non-zero value to avoid divide by zero error
+        f_org = np.where(zero, 0.5, f_org)
+
         # wet radius (m)
         r_wet = ((3 * v_wet) / (4 * np.pi)) ** (1 / 3)
 
-        if f_org == 0:
-            sgm = const.sgm_w
-        else:
-            # C_bulk is the concentration of the organic in the bulk phase
-            # Cb_iso = C_bulk / (1-f_surf)
-            Cb_iso = (f_org * v_dry / const.RUEHL_nu_org) / (v_wet / const.nu_w)
+        # C_bulk is the concentration of the organic in the bulk phase
+        # Cb_iso = C_bulk / (1-f_surf)
+        Cb_iso = (f_org * v_dry / const.RUEHL_nu_org) / (v_wet / const.nu_w)
 
-            # A is the area that one molecule of organic occupies at the droplet surface
-            # A_iso = A*f_surf (m^2)
-            A_iso = (4 * np.pi * r_wet**2) / (
-                f_org * v_dry * const.N_A / const.RUEHL_nu_org
-            )
+        # A is the area that one molecule of organic occupies at the droplet surface
+        # A_iso = A*f_surf (m^2)
+        A_iso = (4 * np.pi * r_wet**2) / (
+            f_org * v_dry * const.N_A / const.RUEHL_nu_org
+        )
 
-            # fraction of organic at surface
-            # quadratic formula, solve equation of state analytically
-            a = -const.RUEHL_A0 / A_iso
-            b = (
-                const.RUEHL_A0 / A_iso
-                + (const.RUEHL_A0 / A_iso) * (const.RUEHL_C0 / Cb_iso)
-                + 1
-            )
-            c = -1
-            f_surf = (-b + np.sqrt(b**2 - 4 * a * c)) / (2 * a)
+        # fraction of organic at surface
+        # quadratic formula, solve equation of state analytically
+        a = -const.RUEHL_A0 / A_iso
+        b = (
+            const.RUEHL_A0 / A_iso
+            + (const.RUEHL_A0 / A_iso) * (const.RUEHL_C0 / Cb_iso)
+            + 1
+        )
+        c = -1
+        f_surf = (-b + np.sqrt(b**2 - 4 * a * c)) / (2 * a)
 
-            # calculate surface tension
-            sgm = const.sgm_w - (
-                (const.R_str * T) / (const.RUEHL_A0 * const.N_A)
-            ) * np.log(1 + Cb_iso * (1 - f_surf) / const.RUEHL_C0)
+        # calculate surface tension
+        _sgm = const.sgm_w - (
+            (const.R_str * T) / (const.RUEHL_A0 * const.N_A)
+        ) * np.log(1 + Cb_iso * (1 - f_surf) / const.RUEHL_C0)
+
+        # where f_org=0 return const.sgm_w, else _sgm
+        _sgm_w = const.sgm_w * np.ones_like(_sgm)
+        sgm = np.where(zero, _sgm_w, _sgm)
 
         # surface tension bounded between sgm_min and sgm_w
         sgm = np.minimum(np.maximum(sgm, const.RUEHL_sgm_min), const.sgm_w)
