### 2022-09-06 1b9fc74
remove if-else so this works for scalars and arrays too
