### 2022-09-08 ca4f1d8
introduce CI steps (codecov, pylint, pdoc and precommit as prerequisites for running long tests)
