+++ .github/workflows/pdoc.yml
@@ -1,33 +1,0 @@
-name: generate and deploy API docs
-
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
-
-jobs:
-  deploy:
-    strategy:
-      matrix:
-        platform: [ubuntu-latest, macos-latest, windows-latest]
-    runs-on: ${{ matrix.platform }}
-    steps:
-      - uses: actions/checkout@v2
-        with:
-          persist-credentials: false
-      - uses: actions/setup-python@v2
-        with:
-          python-version: 3.9
-      - name: Build
-        run: |
-          pip install pdoc3
-          pip install -e .
-          python -We -m pdoc --html PySDM
-      - name: Deploy
-        if: ${{ github.ref == 'refs/heads/main' && matrix.platform == 'ubuntu-latest' }}
-        uses: JamesIves/github-pages-deploy-action@4.1.1
-        with:
-          BRANCH: pdoc
-          FOLDER: html/PySDM
-          CLEAN: true
