+++ .github/workflows/precommithooks.yml
@@ -1,25 +1,0 @@
-name: pre-commit-hooks
-
-on: [pull_request]
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-#         with:
-#           token: ${{ secrets.PYSDM_PRECOMMITHOOK_SECRET }}
-#           ref: ${{ github.head_ref }}
-      - name: Set up Python 3.8
-        uses: actions/setup-python@v1
-        with:
-          python-version: 3.8
-      - name: Linting
-        run: |
-          pip install pre-commit
-          pre-commit clean
-          pre-commit autoupdate
-          pre-commit run --all-files
-#       - name: If needed, commit changes to the pull request
-#         if: failure()
-#         uses: stefanzweifel/git-auto-commit-action@v4
