### 2022-09-13 #4412 @dhhagan
👋 @andreas-h @wholmgren @ethan-nelson @ashwinvis: would you be able to review this manuscript for JOSS?
