### 2022-09-14 17d02e7
Add docstrings and fix strides computation for > 2 dimensions
