+++ PySDM/impl/mesh.py
@@ -32,15 +32,38 @@ class Mesh:
 
     @staticmethod
     def __strides(grid):
+        """
+        returns strides, i.e.: to compute a `cell_id`, use np.dot(strides, cell_origin)
+
+        returns the stride vector for a given grid (for use in `cell_id` arithmetics where
+        the stride vector indicates the distances of cell ids adjacent in a given dimension)
+        """
         domain = np.empty(tuple(grid))
         strides = np.array(domain.strides) // domain.itemsize
         if len(grid) == 1:
             strides = strides.reshape((1, 1))
-        elif len(grid) == 2:
+        else:
             strides = strides.reshape(1, -1)
         return strides
 
     def cellular_attributes(self, positions):
+        """
+        computes values of `cell_id`, `cell_origin` and `position_in_cell` attributes
+        based on `positions` values passed as input
+
+        takes:
+          droplet `positions` as input (expressed in grid coordinates, i.e.
+          in a range of 0 ... 10, 0 ... 5 for a 10x5 grid)
+
+        returns:
+         a tuple of:
+           - `cell_id`: scalar integer ids of cells
+             (each value within the range of 0...prod(grid))
+           - `cell_origin`: n-component vector of integer cell coordinates
+             (each within the range of 0...grid[dim])
+           - `position_in_cell`: n-component vector of floats
+             (each within the range of 0...1)
+        """
         n = positions.shape[1]
         cell_origin = positions.astype(dtype=np.int64)
         position_in_cell = positions - np.floor(positions)
