### 2022-09-14 316902e
Fix: assert cell_origin type equals np int64
