### 2022-09-14 e0b018b
Fix: assert cell_origin type equals np int64
