### 2022-09-15 e234c52
pylint: unnecessary-lambda-assignment
