### 2022-09-15 eda72ff
pylint: duplicate-code
