+++ PySDM/backends/impl_numba/methods/collisions_methods.py
@@ -49,7 +49,7 @@ def coalesce(  # pylint: disable=too-many-arguments
 
 
 @numba.njit(**{**conf.JIT_FLAGS, **{"parallel": False}})
-def break_up(  # pylint: disable=too-many-arguments,unused-argument
+def break_up(  # pylint: disable=too-many-arguments,unused-argument,too-many-locals
     i,
     j,
     k,
@@ -127,7 +127,7 @@ def break_up(  # pylint: disable=too-many-arguments,unused-argument
 
 
 @numba.njit(**{**conf.JIT_FLAGS, **{"parallel": False}})
-def break_up_while(  # pylint: disable=too-many-arguments,unused-argument,too-many-statements
+def break_up_while(  # pylint: disable=too-many-arguments,unused-argument,too-many-statements,too-many-locals
     i,
     j,
     k,
@@ -383,7 +383,7 @@ class CollisionsMethods(BackendMethods):
         volume,
         handle_all_breakups,
     ):
-        # pylint: disable=not-an-iterable,too-many-nested-blocks
+        # pylint: disable=not-an-iterable,too-many-nested-blocks,too-many-locals
         for i in numba.prange(length // 2):
             if gamma[i] == 0:
                 continue
@@ -461,6 +461,7 @@ class CollisionsMethods(BackendMethods):
         volume,
         handle_all_breakups,
     ):
+        # pylint: disable=too-many-locals
         max_multiplicity = np.iinfo(multiplicity.data.dtype).max // 2e5
         self.__collision_coalescence_breakup_body(
             multiplicity=multiplicity.data,
@@ -643,7 +644,7 @@ class CollisionsMethods(BackendMethods):
 
     @staticmethod
     @numba.njit(**conf.JIT_FLAGS)
-    # pylint: disable=too-many-arguments
+    # pylint: disable=too-many-arguments,too-many-locals
     def __compute_gamma_body(
         gamma,
         rand,
