+++ PySDM/backends/impl_numba/methods/condensation_methods.py
@@ -14,7 +14,7 @@ from PySDM.backends.impl_numba.warnings import warn
 
 
 class CondensationMethods(BackendMethods):
-    # pylint: disable=unused-argument
+    # pylint: disable=unused-argument,too-many-locals
     @staticmethod
     def condensation(
         *,
@@ -109,6 +109,7 @@ class CondensationMethods(BackendMethods):
         RH_max,
         success,
     ):
+        # pylint: disable=too-many-locals
         for thread_id in numba.prange(n_threads):  # pylint: disable=not-an-iterable
             for i in range(thread_id, n_cell, n_threads):
                 cell_id = cell_order[i]
@@ -249,7 +250,7 @@ class CondensationMethods(BackendMethods):
         const,
     ):
         @numba.njit(**jit_flags)
-        def step_impl(  # pylint: disable=too-many-arguments
+        def step_impl(  # pylint: disable=too-many-arguments,too-many-locals
             v,
             v_cr,
             n,
@@ -350,7 +351,7 @@ class CondensationMethods(BackendMethods):
         return calculate_ml_old
 
     @staticmethod
-    def make_calculate_ml_new(  # pylint: disable=too-many-statements
+    def make_calculate_ml_new(  # pylint: disable=too-many-statements,too-many-locals
         *,
         jit_flags,
         dx_dt,
@@ -385,7 +386,7 @@ class CondensationMethods(BackendMethods):
             return x_old - x_new + timestep * dx_dt(x_new, r_dr_dt)
 
         @numba.njit(**jit_flags)
-        def calculate_ml_new(  # pylint: disable=too-many-arguments,too-many-statements
+        def calculate_ml_new(  # pylint: disable=too-many-arguments,too-many-statements,too-many-locals
             timestep,
             fake,
             T,
@@ -598,6 +599,7 @@ class CondensationMethods(BackendMethods):
         max_iters,
         const,
     ):
+        # pylint: disable=too-many-locals
         jit_flags = {
             **conf.JIT_FLAGS,
             **{"parallel": False, "cache": False, "fastmath": fastmath},
