+++ tests/unit_tests/dynamics/condensation/test_diagnostics.py
@@ -33,7 +33,7 @@ class _TestEnv:
         return self.full(item)
 
 
-class _TestParticulator:
+class _TestParticulator:  # pylint: disable=too-few-public-methods
     def __init__(  # pylint: disable=too-many-locals
         self,
         *,
@@ -86,7 +86,7 @@ def _try(particulator, capsys):
 FN = fnmatch.translate("*condensation_methods.py")[:-2]
 
 
-class TestDiagnostics:
+class TestDiagnostics:  # pylint: disable=too-few-public-methods
     @staticmethod
     def test_burnout_long(capsys, backend=CPU):
         # arrange
