+++ PySDM/physics/ventilation/neglect.py
@@ -3,6 +3,6 @@ constant ventilation coefficient of unity (i.e., neglect ventilation effects)
 """
 
 
-class Neglect:
+class Neglect:  # pylint: disable=too-few-public-methods
     def __init__(self, _):
         pass
