+++ PySDM/backends/impl_common/random_common.py
@@ -3,7 +3,7 @@ common base class for random number generation abstraction layer
 """
 
 
-class RandomCommon:
+class RandomCommon:  # pylint: disable=too-few-public-methods
     def __init__(self, size: int, seed: int):
         assert isinstance(size, int)
         assert isinstance(seed, int)
