### 2022-09-18 0471b06
pylint: too-many-instance-attributes
