+++ PySDM/backends/impl_thrust_rtc/methods/condensation_methods.py
@@ -12,7 +12,9 @@ from ..conf import trtc
 from ..methods.thrust_rtc_backend_methods import ThrustRTCBackendMethods
 
 
-class CondensationMethods(ThrustRTCBackendMethods):
+class CondensationMethods(
+    ThrustRTCBackendMethods
+):  # pylint: disable=too-many-instance-attributes
     keys = ("T", "p", "pv", "lv", "pvs", "RH", "DTp", "lambdaK", "lambdaD")
 
     def __init__(self):
