+++ PySDM/dynamics/collisions/collision.py
@@ -30,7 +30,7 @@ DEFAULTS = namedtuple("_", ("dt_coal_range", "min_volume", "adaptive", "substeps
 )
 
 
-class Collision:
+class Collision:  # pylint: disable=too-many-instance-attributes
     def __init__(
         self,
         *,
