### 2022-09-23 f27d32f
enable GPU backend tests in test_fragmentations.py
