### 2022-09-23 6879961
Change Attribute's 'size' prop to 'n_vector_components'
