+++ PySDM/attributes/impl/base_attribute.py
@@ -6,8 +6,8 @@ from .attribute import Attribute
 
 
 class BaseAttribute(Attribute):
-    def __init__(self, builder, name, dtype=float, size=0):
-        super().__init__(builder, name=name, dtype=dtype, size=size)
+    def __init__(self, builder, name, dtype=float, n_vector_components=0):
+        super().__init__(builder, name=name, dtype=dtype, n_vector_components=n_vector_components)
 
     def init(self, data):
         self.data.upload(data)
