+++ PySDM/attributes/numerics/position_in_cell.py
@@ -7,5 +7,5 @@ from PySDM.attributes.impl.cell_attribute import CellAttribute
 class PositionInCell(CellAttribute):
     def __init__(self, builder):
         super().__init__(
-            builder, name="position in cell", size=builder.particulator.mesh.dim
+            builder, name="position in cell", n_vector_components=builder.particulator.mesh.dim
         )
