### 2022-09-23 107f5ee
Add 3d displacement with tests, including 1D case
