+++ PySDM/attributes/numerics/cell_origin.py
@@ -7,5 +7,8 @@ from PySDM.attributes.impl.cell_attribute import CellAttribute
 class CellOrigin(CellAttribute):
     def __init__(self, builder):
         super().__init__(
-            builder, name="cell origin", dtype=int, n_vector_components=builder.particulator.mesh.dim
+            builder,
+            name="cell origin",
+            dtype=int,
+            n_vector_components=builder.particulator.mesh.dim,
         )
