+++ tests/unit_tests/dynamics/collisions/test_fragmentations.py
@@ -8,6 +8,7 @@ from PySDM.dynamics.collisions.breakup_fragmentations import (
     SLAMS,
     AlwaysN,
     ExponFrag,
+    Feingold1988Frag,
     Gaussian,
     Straub2010Nf,
 )
@@ -15,7 +16,6 @@ from PySDM.environments import Box
 from PySDM.physics import si
 
 
-# TODO : add tests for the various fragmentation limiters
 class TestFragmentations:  # pylint: disable=too-few-public-methods
     @staticmethod
     @pytest.mark.parametrize(
@@ -23,12 +23,50 @@ class TestFragmentations:  # pylint: disable=too-few-public-methods
         [
             AlwaysN(n=2),
             ExponFrag(scale=1e6 * si.um**3),
+            Feingold1988Frag(scale=1e6 * si.um**3),
             Gaussian(mu=2e6 * si.um**3, sigma=1e6 * si.um**3),
             SLAMS(),
             Straub2010Nf(),
         ],
     )
     def test_fragmentation_fn_call(fragmentation_fn, backend_class=CPU):
+        # arrange
+        volume = np.asarray([440.0 * si.um**3, 6660.0 * si.um**3])
+        fragments = np.asarray([-1.0])
+        builder = Builder(volume.size, backend_class())
+        sut = fragmentation_fn
+        sut.vmin = 1 * si.um**3
+        sut.register(builder)
+        builder.set_environment(Box(dv=None, dt=None))
+        _ = builder.build(attributes={"volume": volume, "n": np.ones_like(volume)})
+
+        _PairwiseStorage = builder.particulator.PairwiseStorage
+        _Indicator = builder.particulator.PairIndicator
+        nf = _PairwiseStorage.from_ndarray(np.zeros_like(fragments))
+        frag_size = _PairwiseStorage.from_ndarray(np.zeros_like(fragments))
+        is_first_in_pair = _Indicator(length=volume.size)
+        u01 = _PairwiseStorage.from_ndarray(np.ones_like(fragments))
+
+        # act
+        sut(nf, frag_size, u01, is_first_in_pair)
+
+        # Assert
+        np.testing.assert_array_less([0.99], nf.to_ndarray())
+        np.testing.assert_array_less([0.0], frag_size.to_ndarray())
+
+    @staticmethod
+    @pytest.mark.parametrize(
+        "fragmentation_fn",
+        [
+            ExponFrag(scale=1 * si.um**3, vmin=6660.0 * si.um**3),
+            Feingold1988Frag(scale=1 * si.um**3, vmin=6660.0 * si.um**3),
+            Gaussian(mu=2 * si.um**3, sigma=1 * si.um**3, vmin=6660.0 * si.um**3),
+            SLAMS(vmin=6660.0 * si.um**3),
+            Straub2010Nf(vmin=6660.0 * si.um**3),
+            pytest.param(AlwaysN(n=10), marks=pytest.mark.xfail(strict=True)),
+        ],
+    )
+    def test_fragmentation_limiters_vmin(fragmentation_fn, backend_class=CPU):
         # arrange
         volume = np.asarray([440.0 * si.um**3, 6660.0 * si.um**3])
         fragments = np.asarray([-1.0])
@@ -51,5 +89,81 @@ class TestFragmentations:  # pylint: disable=too-few-public-methods
         print(nf.data, frag_size.data)
 
         # Assert
-        np.testing.assert_array_less([0.99], nf.to_ndarray())
-        np.testing.assert_array_less([0.0], frag_size.to_ndarray())
+        np.testing.assert_array_equal([(440.0 + 6660.0) / 6660.0], nf.to_ndarray())
[…]
