### 2022-09-26 64f5cb9
add test for collision efficiencies, cleanup AlwaysN
