+++ tests/smoke_tests/lowe_et_al_2019/test_zero_forg.py
@@ -35,10 +35,10 @@ def test_zero_forg(plot=False):  # pylint: disable=too-many-locals
                     n_sd_per_mode=20,
                     model=model,
                     aerosol={
-                        "a": AerosolMarine(Forg=0, Acc_N2=Acc["a"]),
-                        "b": AerosolMarine(Forg=0, Acc_N2=Acc["b"]),
-                        "c": AerosolBoreal(Forg=0, Acc_N2=Acc["c"]),
-                        "d": AerosolBoreal(Forg=0, Acc_N2=Acc["d"]),
+                        "a": AerosolMarine(Acc_N2=Acc["a"]),
+                        "b": AerosolMarine(Acc_N2=Acc["b"]),
+                        "c": AerosolBoreal(Acc_N2=Acc["c"]),
+                        "d": AerosolBoreal(Acc_N2=Acc["d"]),
                     }[subplot],
                     w=w * si.m / si.s,
                     spectral_sampling=spec_sampling.ConstantMultiplicity,
