### 2022-09-27 df251da
fix GPU code compilation of RhodThdQv::T
