### 2022-09-27 af51028
add test for c_inline (checking the new trailing come avoidance logic)
