+++ PySDM/backends/impl_thrust_rtc/conf.py
@@ -3,13 +3,15 @@ ThrustRTC import logic implementing the switch between the genuine ThrustRTC and
 default nice_thrust flags
 """
 from warnings import warn
-
+import os
 from PySDM.backends.impl_thrust_rtc.test_helpers.flag import fakeThrustRTC
 
 if not fakeThrustRTC:
     try:
         # noinspection PyUnresolvedReferences
         import ThrustRTC as trtc  # pylint: disable=unused-import
+        if 'NVRTC_PATH' in os.environ:
+            trtc.set_libnvrtc_path(os.environ['NVRTC_PATH'])
     except OSError as error:
         warn(f"importing ThrustRTC failed with {error}")
     try:
