### 2022-09-29 30c0cf7
Remove multicell rate test
