### 2022-09-29 cf828ad
Clean up testing file
