+++ tests/unit_tests/dynamics/collisions/test_fragmentations.py
@@ -86,8 +86,6 @@ class TestFragmentations:  # pylint: disable=too-few-public-methods
         # act
         sut(nf, frag_size, u01, is_first_in_pair)
 
-        print(nf.data, frag_size.data)
-
         # Assert
         np.testing.assert_array_equal([(440.0 + 6660.0) / 6660.0], nf.to_ndarray())
         np.testing.assert_array_equal([6660.0 * si.um**3], frag_size.to_ndarray())
