### 2022-10-01 e978b11
add support for vectorizing formulae (using either np.vectorize or numba.vectorize). Closes #748.
