### 2022-10-02 ca7d5bd
move numba-experimental-feature wrning filtering into Formula; vectorize Szyszkowski-Langmuir sigma; cleanup surface tension tests from py_func usage
