### 2022-10-05 3770215
fix EffectiveRadius product so its register() method can be called twice
