### 2022-10-07 d440eba
Small refactor
