+++ tests/unit_tests/dynamics/displacement/test_advection.py
@@ -11,10 +11,14 @@ assert hasattr(backend_class, "_pytestfixturefunction")
 class TestExplicitEulerWithInterpolation:
     @staticmethod
     @pytest.mark.parametrize(
-        "positions, courant_field",
+        "positions, courant_field, expected_positions",
         (
             # 1D
-            ([[0.5]], (np.array([0.1, 0.2]),)),
+            (
+                [[0.5]],
+                (np.array([0.1, 0.2]),),
+                [[0.65827668]],
+            ),
             # 2D
             (
                 [[0.5], [0.5]],
@@ -22,6 +26,7 @@ class TestExplicitEulerWithInterpolation:
                     np.array([0.1, 0.2]).reshape((2, 1)),
                     np.array([0.3, 0.4]).reshape((1, 2)),
                 ),
+                [[0.65827668], [0.86931224]],
             ),
             # 3D
             (
@@ -29,13 +34,18 @@ class TestExplicitEulerWithInterpolation:
                 (
                     np.array([0.1, 0.2]).reshape((2, 1, 1)),
                     np.array([0.3, 0.4]).reshape((1, 2, 1)),
-                    np.array([0.5, 0.6]).reshape((1, 1, 2)),
+                    np.array([0.2, 0.3]).reshape((1, 1, 2)),
                 ),
+                [
+                    [0.65827668],
+                    [0.86931224],
+                    [0.76379446],
+                ]
             ),
         ),
     )
     # pylint: disable=redefined-outer-name
-    def test_single_cell(backend_class, positions, courant_field: tuple):
+    def test_single_cell(backend_class, positions, expected_positions, courant_field: tuple):
         # Arrange
         settings = DisplacementSettings(
             courant_field_data=courant_field,
@@ -43,13 +53,16 @@ class TestExplicitEulerWithInterpolation:
             grid=tuple([1] * len(courant_field)),
             n_sd=len(positions[0]),
         )
-        sut, _ = settings.get_displacement(backend_class, scheme="ImplicitInSpace")
+        sut, particulator = settings.get_displacement(backend_class, scheme="ImplicitInSpace")
 
         # Act
         sut()
 
         # Assert
-        print(positions)
+        np.testing.assert_array_almost_equal(
+            np.asarray(expected_positions),
+            particulator.attributes['position in cell'].to_ndarray()
+        )
 
     @staticmethod
     # pylint: disable=redefined-outer-name
