### 2022-10-07 5348730
pylint suggestions (incl. FakeThrust API arg names sync with Thrust)
