+++ PySDM/backends/impl_thrust_rtc/methods/displacement_methods.py
@@ -12,7 +12,7 @@ class DisplacementMethods(ThrustRTCBackendMethods):
     def __init__(self):
         super().__init__()
         self.__calculate_displacement_body = trtc.For(
-            args=(
+            param_names=(
                 "dim",
                 "n_sd",
                 "displacement",
@@ -23,10 +23,10 @@ class DisplacementMethods(ThrustRTCBackendMethods):
                 "position_in_cell",
                 "n_substeps",
             ),
-            iter_var="i",
+            name_iter="i",
             body=f"""
             // Arakawa-C grid
-            
+
             // 1D, 2D & 3D
             auto _l = cell_origin[i];
             auto _r = cell_origin[i] + 1 * (dim == 0);
@@ -39,7 +39,7 @@ class DisplacementMethods(ThrustRTCBackendMethods):
                 _l += _l_1 * courant_shape_0;
                 _r += _r_1 * courant_shape_0;
             }}
-            
+
             // 3D
             if (dim == 2) {{
                 auto _l_1 = cell_origin[i + n_sd];
@@ -47,11 +47,11 @@ class DisplacementMethods(ThrustRTCBackendMethods):
 
                 auto _l_2 = cell_origin[i + 2 * n_sd];
                 auto _r_2 = cell_origin[i + 2 * n_sd] + 1 * (dim == 2);
-                
+
                 _l += _l_1 * courant_shape_0 + _l_2 * courant_shape_0 * courant_shape_1
                 _r += _r_1 * courant_shape_0 + _r_2 * courant_shape_0 * courant_shape_1
             }}
-            
+
             auto omega = position_in_cell[i + n_sd * dim];
             auto c_r = courant[_r] / n_substeps;
             auto c_l = courant[_l] / n_substeps;
@@ -98,9 +98,11 @@ class DisplacementMethods(ThrustRTCBackendMethods):
         dim = trtc.DVInt64(dim)
         n_sd = trtc.DVInt64(position_in_cell.shape[1])
         courant_shape_0 = trtc.DVInt64(courant.shape[0])
-        courant_shape_1 = trtc.DVInt64(courant.shape[1] if len(courant.shape) > 2 else -1)
+        courant_shape_1 = trtc.DVInt64(
+            courant.shape[1] if len(courant.shape) > 2 else -1
+        )
         self.__calculate_displacement_body.launch_n(
-            size=displacement.shape[1],
+            n=displacement.shape[1],
             args=(
                 dim,
                 n_sd,
