### 2022-10-07 aa29bcc
Add exp_fragmentation_body for ThrustRTC
