### 2022-10-07 0ad1dee
fix missing semicolons in C++ code
