### 2022-10-07 252025e
ZERO_DIM_MESH = (1,)
