### 2022-10-08 fc34595
Thrust C++ code refactor to speed up 1D and 2D displacement (number of dimensions as compile-time constants)
