### 2022-10-08 ec5d30b
Mesh code cleanup
