### 2022-10-14 7eab263
Fix test_fragmentation_fn_call and add Gaussian and SLAMS fragmentation methods for ThrustRTC
