### 2022-10-14 20dd17c
Sync FakeThrust API arg names with Thrust
