### 2022-10-14 8485938
Add native implementation for addition on ThrustRTC (built-in Plus not working for different arg types)
