### 2022-10-14 defc707
Import CPU to test_fragmentations
