### 2022-10-14 6a86b4f
making it possible to call Builder::request_attribute before Builder::build. Closes #751.
