### 2022-10-14 658077d
test data: avoid divide-by-zero warning when run without Numba
