### 2022-10-14 085cddc
Add the option to remove a dynamic from builder
