### 2022-10-19 845a173
GPU freezing methods
