+++ tests/unit_tests/backends/test_freezing_methods.py
@@ -1,44 +1,78 @@
 # pylint: disable=missing-module-docstring,missing-class-docstring,missing-function-docstring
 import numpy as np
-from matplotlib import pylab
+from matplotlib import pyplot
 
 from PySDM import Builder, Formulae
-from PySDM.backends import CPU
 from PySDM.dynamics import Freezing
 from PySDM.environments import Box
 from PySDM.physics import constants_defaults as const
+from PySDM.physics import si
 from PySDM.products import IceWaterContent
 
-from ...backends_fixture import backend_class  # TODO #599
+from ...backends_fixture import backend_class
 
 assert hasattr(backend_class, "_pytestfixturefunction")
 
 
 class TestFreezingMethods:
     # TODO #599
-    @staticmethod
-    # pylint: disable=redefined-outer-name
-    def test_freeze_singular(backend_class):
+    def test_subsaturated_freezing(self):
         pass
 
     @staticmethod
-    # pylint: disable=too-many-locals
-    def test_freeze_time_dependent(plot=False):
+    # pylint: disable=redefined-outer-name
+    def test_freeze_singular(backend_class):
+        # arrange
+        n_sd = 44
+        dt = 1 * si.s
+        dv = 1 * si.m**3
+        T_fz = 250 * si.K
+        vol = 1 * si.um**3
+        multiplicity = 1e10
+        steps = 1
+
+        formulae = Formulae()
+        builder = Builder(n_sd=n_sd, backend=backend_class(formulae=formulae))
+        env = Box(dt=dt, dv=dv)
+        builder.set_environment(env)
+        builder.add_dynamic(Freezing(singular=True))
+        attributes = {
+            "n": np.full(n_sd, multiplicity),
+            "freezing temperature": np.full(n_sd, T_fz),
+            "volume": np.full(n_sd, vol),
+        }
+        products = (IceWaterContent(name="qi"),)
+        particulator = builder.build(attributes=attributes, products=products)
+        env["T"] = T_fz
+        env["RH"] = 1.000001
+
+        # act
+        particulator.run(steps=steps)
+
+        # assert
+        np.testing.assert_almost_equal(
+            np.asarray(particulator.products["qi"].get()),
+            [n_sd * multiplicity * vol * const.rho_w / dv],
+        )
+
+    @staticmethod
+    # pylint: disable=too-many-locals,redefined-outer-name
+    def test_freeze_time_dependent(backend_class, plot=False):
         # Arrange
         cases = (
-            {"dt": 5e5, "N": 1},
-            {"dt": 1e6, "N": 1},
-            {"dt": 5e5, "N": 8},
-            {"dt": 1e6, "N": 8},
-            {"dt": 5e5, "N": 16},
-            {"dt": 1e6, "N": 16},
+            {"dt": 5e4, "N": 1},
+            {"dt": 1e5, "N": 1},
+            {"dt": 5e4, "N": 8},
+            {"dt": 1e5, "N": 8},
+            {"dt": 5e4, "N": 16},
+            {"dt": 1e5, "N": 16},
         )
         rate = 1e-9
         immersed_surface_area = 1
 
         number_of_real_droplets = 1024
         total_time = (
-            2e9  # effectively interpretted here as seconds, i.e. cycle = 1 * si.s
+            0.25e9  # effectively interpreted here as seconds, i.e. cycle = 1 * si.s
         )
 
         # dummy (but must-be-set) values
@@ -48,10 +82,10 @@ class TestFreezingMethods:
         d_v = 666  # products use conc., dividing there, multiplying here, value does not matter
 
         def hgh(t):
[…]
