### 2022-10-19 275843d
fix FakeThrust compatibility
