+++ PySDM/formulae.py
@@ -41,15 +41,12 @@ class Formulae:  # pylint: disable=too-few-public-methods,too-many-instance-attr
         freezing_temperature_spectrum: str = "Null",
         heterogeneous_ice_nucleation_rate: str = "Null",
     ):
-        constants_defaults = {}
-        for defaults in (physics.constants, physics.constants_defaults):
-            constants_defaults.update(
-                {
-                    k: getattr(defaults, k)
-                    for k in dir(defaults)
-                    if isinstance(getattr(defaults, k), (numbers.Number, pint.Quantity))
-                }
-            )
+        constants_defaults = {
+            k: getattr(defaults, k)
+            for defaults in (physics.constants, physics.constants_defaults)
+            for k in dir(defaults)
+            if isinstance(getattr(defaults, k), (numbers.Number, pint.Quantity))
+        }
         constants = namedtuple("Constants", tuple(constants_defaults.keys()))(
             **{**constants_defaults, **(constants or {})}
         )
