### 2022-10-19 df5648e
shorten and fix random seed for the freezing test
