### 2022-10-20 b363b7f
Add min_pair, multiply_pair and div_if_not_0 implementation for ThrustRTC
