+++ PySDM/backends/impl_thrust_rtc/methods/pair_methods.py
@@ -138,3 +138,61 @@ class PairMethods(ThrustRTCBackendMethods):
             n=len(idx),
             args=(data_out.data, perm_in, is_first_in_pair.indicator.data),
         )
+
+    __min_pair_body = trtc.For(
+        param_names=(
+            "data_out",
+            "data_in",
+            "is_first_in_pair",
+            "idx",
+        ),
+        name_iter="i",
+        body="""
+        if (is_first_in_pair[i]) {
+            data_out[(int64_t)(i/2)] = min(data_in[idx[i]], data_in[idx[i + 1]]);
+        }
+        """,
+    )
+
+    @staticmethod
+    @nice_thrust(**NICE_THRUST_FLAGS)
+    def min_pair(data_out, data_in, is_first_in_pair, idx):
+        trtc.Fill(data_out.data, trtc.DVDouble(0))
+        PairMethods.__min_pair_body.launch_n(
+            n=len(idx),
+            args=(
+                data_out.data,
+                data_in.data,
+                is_first_in_pair.indicator.data,
+                idx.data,
+            ),
+        )
+
+    __multiply_pair_body = trtc.For(
+        param_names=(
+            "data_out",
+            "data_in",
+            "is_first_in_pair",
+            "idx",
+        ),
+        name_iter="i",
+        body="""
+        if (is_first_in_pair[i]) {
+            data_out[(int64_t)(i/2)] = data_in[idx[i]] * data_in[idx[i + 1]];
+        }
+        """,
+    )
+
+    @staticmethod
+    @nice_thrust(**NICE_THRUST_FLAGS)
+    def multiply_pair(data_out, data_in, is_first_in_pair, idx):
+        trtc.Fill(data_out.data, trtc.DVDouble(0))
+        PairMethods.__multiply_pair_body.launch_n(
+            n=len(idx),
+            args=(
+                data_out.data,
+                data_in.data,
+                is_first_in_pair.indicator.data,
+                idx.data,
+            ),
+        )
