### 2022-10-20 3879f7e
Add implementation of fragmentation functions for ThrustRTC
