+++ PySDM/backends/impl_thrust_rtc/methods/collisions_methods.py
@@ -3,7 +3,6 @@ GPU implementation of backend methods for particle collisions
 """
 from PySDM.backends.impl_thrust_rtc.conf import NICE_THRUST_FLAGS
 from PySDM.backends.impl_thrust_rtc.nice_thrust import nice_thrust
-from PySDM.physics.constants import sqrt_pi, sqrt_two
 
 from ..conf import trtc
 from ..methods.thrust_rtc_backend_methods import ThrustRTCBackendMethods
@@ -14,6 +13,7 @@ class CollisionsMethods(
 ):  # pylint: disable=too-many-instance-attributes
     def __init__(self):
         super().__init__()
+        const = self.formulae.constants
 
         self.__adaptive_sdm_gamma_body_1 = trtc.For(
             ("dt_todo", "dt_left", "dt_max"),
@@ -189,10 +189,10 @@ class CollisionsMethods(
         )
 
         self.__exp_fragmentation_body = trtc.For(
-            param_names=("scale", "frag_size", "rand"),
+            param_names=("scale", "frag_size", "rand, tol"),
             name_iter="i",
             body="""
-            frag_size[i] = -scale * log(1 - rand[i]);
+            frag_size[i] = -scale * log(max(1 - rand[i], tol));
             """,
         )
 
@@ -203,35 +203,39 @@ class CollisionsMethods(
                 "v_max",
                 "x_plus_y",
                 "vmin",
-                "nfmax_is_not_none",
                 "nfmax",
+                "nfmax_is_not_none",
             ),
             name_iter="i",
             body="""
-            n_fragment[i] = x_plus_y[i] / frag_size[i];
-            if (frag_size[i] > v_max[i]) {
-                n_fragment[i] = 1;
-            }
-            else if (frag_size[i] < vmin) {
-                n_fragment[i] = 1;
-            }
+            frag_size[i] = min(frag_size[i], v_max[i]);
+            frag_size[i] = max(frag_size[i], vmin);
 
             if (nfmax_is_not_none) {
-                n_fragment[i] = min(n_fragment[i], nfmax) / 0;
+                if (x_plus_y[i] / frag_size[i] > nfmax) {
+                    frag_size[i] = x_plus_y[i] / nfmax;
+                }
             }
+            if (frag_size[i] == 0.0) {
+                frag_size[i] = x_plus_y[i];
+                n_fragment[i] = 1.0;
+            }
+            n_fragment[i] = x_plus_y[i] / frag_size[i];
             """,
         )
 
         self.__gauss_fragmentation_body = trtc.For(
-            param_names=("mu", "sigma", "frag_size", "rand", "sqrt_pi", "sqrt_two"),
+            param_names=("mu", "sigma", "frag_size", "rand"),
             name_iter="i",
-            body="""
-            frag_size[i] = mu + sqrt_pi * sqrt_two * sigma / 4 * log((1 + rand[i]) / (1 - rand[i]));
+            body=f"""
+            frag_size[i] = mu - sigma / {const.sqrt_two} / {const.sqrt_pi} / log(2) * log(
+                (0.5 + rand[i]) / (1.5 - rand[i])
+            );
             """,
         )
 
         self.__slams_fragmentation_body = trtc.For(
-            param_names=("n_fragment", "probs", "rand"),
+            param_names=("n_fragment", "frag_size", "x_plus_y", "probs", "rand"),
             name_iter="i",
             body="""
             probs[i] = 0.0;
@@ -244,6 +248,135 @@ class CollisionsMethods(
                     break;
                 }
             }
+            frag_size[i] = x_plus_y[i] / n_fragment[i];
+            """,
+        )
+
[…]
