+++ tests/unit_tests/dynamics/collisions/test_fragmentations.py
@@ -75,7 +75,7 @@ class TestFragmentations:  # pylint: disable=too-few-public-methods
             pytest.param(AlwaysN(n=10), marks=pytest.mark.xfail(strict=True)),
         ],
     )
-    def test_fragmentation_limiters_vmin(fragmentation_fn, backend_class=CPU):
+    def test_fragmentation_limiters_vmin(fragmentation_fn, backend_class):
         # arrange
         volume = np.asarray([440.0 * si.um**3, 6660.0 * si.um**3])
         fragments = np.asarray([-1.0])
@@ -90,6 +90,9 @@ class TestFragmentations:  # pylint: disable=too-few-public-methods
         nf = _PairwiseStorage.from_ndarray(np.zeros_like(fragments))
         frag_size = _PairwiseStorage.from_ndarray(np.zeros_like(fragments))
         is_first_in_pair = _Indicator(length=volume.size)
+        is_first_in_pair.indicator = builder.particulator.Storage.from_ndarray(
+            np.asarray([True, False])
+        )
         u01 = _PairwiseStorage.from_ndarray(np.ones_like(fragments))
 
         # act
@@ -127,6 +130,9 @@ class TestFragmentations:  # pylint: disable=too-few-public-methods
         nf = _PairwiseStorage.from_ndarray(np.zeros_like(fragments))
         frag_size = _PairwiseStorage.from_ndarray(np.zeros_like(fragments))
         is_first_in_pair = _Indicator(length=volume.size)
+        is_first_in_pair.indicator = builder.particulator.Storage.from_ndarray(
+            np.asarray([True, False])
+        )
         u01 = _PairwiseStorage.from_ndarray(np.ones_like(fragments))
 
         # act
@@ -164,6 +170,9 @@ class TestFragmentations:  # pylint: disable=too-few-public-methods
         nf = _PairwiseStorage.from_ndarray(np.zeros_like(fragments))
         frag_size = _PairwiseStorage.from_ndarray(np.zeros_like(fragments))
         is_first_in_pair = _Indicator(length=volume.size)
+        is_first_in_pair.indicator = builder.particulator.Storage.from_ndarray(
+            np.asarray([True, False])
+        )
         u01 = _PairwiseStorage.from_ndarray(np.ones_like(fragments))
 
         # act
