+++ tests/unit_tests/dynamics/collisions/test_fragmentations.py
@@ -3,7 +3,6 @@ import numpy as np
 import pytest
 
 from PySDM import Builder
-from PySDM.backends import CPU
 from PySDM.dynamics.collisions.breakup_fragmentations import (
     SLAMS,
     AlwaysN,
@@ -79,7 +78,7 @@ class TestFragmentations:  # pylint: disable=too-few-public-methods
         # arrange
         volume = np.asarray([440.0 * si.um**3, 6660.0 * si.um**3])
         fragments = np.asarray([-1.0])
-        builder = Builder(volume.size, backend_class())
+        builder = Builder(volume.size, backend_class(double_precision=True))
         sut = fragmentation_fn
         sut.register(builder)
         builder.set_environment(Box(dv=None, dt=None))
@@ -114,7 +113,7 @@ class TestFragmentations:  # pylint: disable=too-few-public-methods
             pytest.param(AlwaysN(n=0.01), marks=pytest.mark.xfail(strict=True)),
         ],
     )
-    def test_fragmentation_limiters_vmax(fragmentation_fn, backend_class=CPU):
+    def test_fragmentation_limiters_vmax(fragmentation_fn, backend_class):
         # arrange
         volume = np.asarray([440.0 * si.um**3, 6660.0 * si.um**3])
         fragments = np.asarray([-1.0])
@@ -154,7 +153,7 @@ class TestFragmentations:  # pylint: disable=too-few-public-methods
             pytest.param(AlwaysN(n=10), marks=pytest.mark.xfail(strict=True)),
         ],
     )
-    def test_fragmentation_limiters_nfmax(fragmentation_fn, backend_class=CPU):
+    def test_fragmentation_limiters_nfmax(fragmentation_fn, backend_class):
         # arrange
         volume = np.asarray([440.0 * si.um**3, 6660.0 * si.um**3])
         fragments = np.asarray([-1.0])
