+++ tests/unit_tests/dynamics/collisions/test_fragmentations.py
@@ -74,7 +74,9 @@ class TestFragmentations:  # pylint: disable=too-few-public-methods
             pytest.param(AlwaysN(n=10), marks=pytest.mark.xfail(strict=True)),
         ],
     )
-    def test_fragmentation_limiters_vmin(fragmentation_fn, backend_class):
+    def test_fragmentation_limiters_vmin(
+        fragmentation_fn, backend_class
+    ):  # pylint: disable=redefined-outer-name
         # arrange
         volume = np.asarray([440.0 * si.um**3, 6660.0 * si.um**3])
         fragments = np.asarray([-1.0])
@@ -113,7 +115,9 @@ class TestFragmentations:  # pylint: disable=too-few-public-methods
             pytest.param(AlwaysN(n=0.01), marks=pytest.mark.xfail(strict=True)),
         ],
     )
-    def test_fragmentation_limiters_vmax(fragmentation_fn, backend_class):
+    def test_fragmentation_limiters_vmax(
+        fragmentation_fn, backend_class
+    ):  # pylint: disable=redefined-outer-name
         # arrange
         volume = np.asarray([440.0 * si.um**3, 6660.0 * si.um**3])
         fragments = np.asarray([-1.0])
@@ -153,7 +157,9 @@ class TestFragmentations:  # pylint: disable=too-few-public-methods
             pytest.param(AlwaysN(n=10), marks=pytest.mark.xfail(strict=True)),
         ],
     )
-    def test_fragmentation_limiters_nfmax(fragmentation_fn, backend_class):
+    def test_fragmentation_limiters_nfmax(
+        fragmentation_fn, backend_class
+    ):  # pylint: disable=redefined-outer-name
         # arrange
         volume = np.asarray([440.0 * si.um**3, 6660.0 * si.um**3])
         fragments = np.asarray([-1.0])
