### 2022-10-20 116944b
Fix for loop in slams thrustrtc impl
