### 2022-10-20 dd35cc1
Reuse key-generation and add simple unit test
