### 2022-10-20 4c5cb8c
rename test, avoid double instantiation, check if removed in assert block
