### 2022-10-20 1c295be
assert on object identity in remove_dynamic() to avoid misleading calling sequences
