+++ PySDM/builder.py
@@ -45,12 +45,11 @@ class Builder:
         assert key not in self.particulator.dynamics
         self.particulator.dynamics[key] = dynamic
 
-    def remove_dynamic(self, dynamic):
-        assert self.particulator.environment is not None
+    def replace_dynamic(self, dynamic):
         key = get_key(dynamic)
         assert key in self.particulator.dynamics
-        assert dynamic is self.particulator.dynamics[key]
         self.particulator.dynamics.pop(key)
+        self.add_dynamic(dynamic)
 
     def register_product(self, product, buffer):
         if product.name in self.particulator.products:
