### 2022-10-21 3570245
speeding up the BDF solver by removing numba decorators - do not play well with SciPy (+disabling warning catchers)
