+++ PySDM/backends/impl_numba/test_helpers/bdf.py
@@ -3,7 +3,6 @@ condensation/evaporation solver drop-in replacement implemented using
  SciPy adaptive-timestep BDF solver, for use in tests only
 """
 import types
-import warnings
 from functools import lru_cache
 
 import numba
@@ -88,7 +87,6 @@ def _make_solve(formulae):  # pylint: disable=too-many-statements,too-many-local
     def _ql(n, x, m_d_mean):
         return np.sum(n * volume(x)) * rho_w / m_d_mean
 
-    @numba.njit(**{**JIT_FLAGS, **{"parallel": False}})
     def _impl(  # pylint: disable=too-many-arguments,too-many-locals
         dy_dt,
         x,
@@ -132,7 +130,6 @@ def _make_solve(formulae):  # pylint: disable=too-many-statements,too-many-local
         dqv_dt = dot_qv - np.sum(n * volume(x) * dy_dt[idx_x:]) * rho_w / m_d_mean
         dy_dt[idx_thd] = dot_thd + phys_dthd_dt(rhod_mean, thd, T, dqv_dt, lv)
 
-    @numba.njit(**{**JIT_FLAGS, **{"parallel": False}})
     def _odesys(  # pylint: disable=too-many-arguments,too-many-locals
         t, y, kappa, f_org, dry_volume, n, dthd_dt, dqv_dt, m_d_mean, rhod_mean, qt
     ):
@@ -205,18 +202,16 @@ def _make_solve(formulae):  # pylint: disable=too-many-statements,too-many-local
         if dthd_dt == 0 and dqv_dt == 0 and (_odesys(0, y0, *args)[idx_x] == 0).all():
             y1 = y0
         else:
-            with warnings.catch_warnings(record=True) as _:
-                warnings.simplefilter("ignore")
-                integ = scipy.integrate.solve_ivp(
-                    fun=_odesys,
-                    args=args,
-                    t_span=(0, dt),
-                    t_eval=(dt,),
-                    y0=y0,
-                    rtol=rtol,
-                    atol=0,
-                    method="BDF",
-                )
+            integ = scipy.integrate.solve_ivp(
+                fun=_odesys,
+                args=args,
+                t_span=(0, dt),
+                t_eval=(dt,),
+                y0=y0,
+                rtol=rtol,
+                atol=0,
+                method="BDF",
+            )
             assert integ.success, integ.message
             y1 = integ.y[:, 0]
 
