### 2022-10-22 37d673c
handling ql->qc product name change
