### 2022-10-22 9953810
removing " rate" suffix from produc names
