+++ PySDM/backends/impl_numba/methods/collisions_methods.py
@@ -291,7 +291,7 @@ def straub_p3(  # pylint: disable=too-many-arguments,unused-argument
     rand,
 ):
     mu3 = 0.9 * ds[i]
-    delD3 = 0.01 * (0.76 * CW[i] ** (1 / 2) + 1.0)
+    delD3 = 0.01 * (0.76 * CW[i] ** 1 / 2 + 1.0)
     sigma3 = delD3**2 / 12
     X = rand[i]
 
