### 2022-10-23 aac0d1e
fix typo in Straub et al. formulae (thanks @abulenok!) (#934)
