### 2022-10-23 5e2b5f9
add unit tests for straub parameter fixes
