### 2022-10-23 8480c6c
Fix a few typos in straub fragmentation backend (#937)
Co-authored-by: Sylwester Arabas <sylwester.arabas@uj.edu.pl>
