### 2022-10-23 52e6f1c
Fix typos in Straub fragmentation function
