### 2022-10-23 628c29a
Test only CPU version of sum_pair_body for out of bounds exception (no numba)
