### 2022-10-23 1120a5a
Add test for sum_pair for CPU and GPU
