+++ tests/unit_tests/backends/test_pair_methods.py
@@ -1,9 +1,15 @@
 # pylint: disable=missing-module-docstring,missing-class-docstring,missing-function-docstring
+import os
+
 import numpy as np
 import pytest
 
 from PySDM.backends import CPU
 from PySDM.backends.impl_common.pair_indicator import make_PairIndicator
+
+from ...backends_fixture import backend_class
+
+assert hasattr(backend_class, "_pytestfixturefunction")
 
 
 @pytest.mark.parametrize(
@@ -45,9 +51,49 @@ def test_sum_pair_body_out_of_bounds(
 
     idx = backend.Storage.from_ndarray(np.asarray(_idx))
 
+    sut = (
+        backend.sum_pair_body
+        if "NUMBA_DISABLE_JIT" in os.environ
+        else backend.sum_pair_body.py_func
+    )
     # Act
-    backend.sum_pair_body.py_func(
+    sut(
         data_out.data, data_in.data, is_first_in_pair.indicator.data, idx.data, len(idx)
     )
 
     # Assert
+
+
+@pytest.mark.parametrize(
+    "_data_in, _data_out, _idx",
+    (
+        pytest.param(
+            [44.0, 666.0],
+            [
+                0,
+            ],
+            [0, 1],
+        ),
+    ),
+)
+def test_sum_pair(_data_in, _data_out, _idx, backend_class):
+    # Arrange
+    backend = backend_class()
+
+    data_out = backend.Storage.from_ndarray(np.asarray(_data_out))
+    data_in = backend.Storage.from_ndarray(np.asarray(_data_in))
+
+    is_first_in_pair = make_PairIndicator(backend)(len(_data_in))
+    is_first_in_pair.indicator = backend.Storage.from_ndarray(
+        np.asarray(
+            [True, False],
+        )
+    )
+
+    idx = backend.Storage.from_ndarray(np.asarray(_idx))
+
+    # Act
+    backend.sum_pair(data_out, data_in, is_first_in_pair, idx)
+
+    # Assert
+    np.testing.assert_array_equal(data_out, [44.0 + 666.0])
