### 2022-10-24 80da6f8
Add implicit type casting to make CUDA pow func work
