+++ PySDM/backends/impl_thrust_rtc/methods/collisions_methods.py
@@ -282,66 +282,74 @@ class CollisionsMethods(
         self.__straub_p1 = f"""
 
                 auto E_D1 = 0.04 * CM;
-                auto delD1 = 0.0125 * pow(CW[i], 0.5);
+                auto delD1 = 0.0125 * pow(CW[i], (real_type) (0.5));
                 auto var_1 = pow(delD1, 2.) / 12.;
-                auto sigma1 = sqrt(log(var_1 / pow(E_D1, 2) + 1));
-                auto mu1 = log(E_D1) - pow(sigma1, 2) / 2.;
+                auto sigma1 = sqrt(log(var_1 / pow(E_D1, 2.) + 1));
+                auto mu1 = log(E_D1) - pow(sigma1, 2.) / 2.;
                 auto X = rand[i];
 
                 frag_size[i] = exp(
                     mu1
                     - sigma1 / {const.sqrt_two} / {const.sqrt_pi} / log(2.) * log((0.5 + X) / (1.5 - X))
                 );
-                frag_size[i] = {const.PI} / 6. * pow(frag_size[i], 3);
-        """
+                frag_size[i] = {const.PI} / 6. * pow(frag_size[i], (real_type) (3.));
+        """.replace(
+            "real_type", self._get_c_type()
+        )
 
         self.__straub_p2 = f"""
 
                 auto mu2 = 0.095 * CM;
                 auto delD2 = 0.007 * (CW[i] - 21.0);
-                auto sigma2 = pow(delD2, 2) / 12.;
+                auto sigma2 = pow(delD2, 2.) / 12.;
                 auto X = rand[i];
 
                 frag_size[i] = mu2 - sigma2 / {const.sqrt_two} / {const.sqrt_pi} / log(2.) * log(
                     (0.5 + X) / (1.5 - X)
                 );
-                frag_size[i] = {const.PI} / 6. * pow(frag_size[i], 3);
-        """
+                frag_size[i] = {const.PI} / 6. * pow(frag_size[i], (real_type) (3.));
+        """.replace(
+            "real_type", self._get_c_type()
+        )
 
         self.__straub_p3 = f"""
 
                 auto mu3 = 0.9 * ds[i];
-                auto delD3 = 0.01 * (0.76 * pow(CW[i], 0.5) + 1.0);
-                auto sigma3 = pow(delD3, 2) / 12.;
+                auto delD3 = 0.01 * (0.76 * pow(CW[i], (real_type) (0.5)) + 1.0);
+                auto sigma3 = pow(delD3, 2.) / 12.;
                 auto X = rand[i];
 
                 frag_size[i] = mu3 - sigma3 / {const.sqrt_two} / {const.sqrt_pi} / log(2.) * log(
                     (0.5 + X) / (1.5 - X)
                 );
-                frag_size[i] = {const.PI} / 6. * pow(frag_size[i], 3);
-        """
+                frag_size[i] = {const.PI} / 6. * pow(frag_size[i], (real_type) (3.));
+        """.replace(
+            "real_type", self._get_c_type()
+        )
 
         self.__straub_p4 = f"""
 
                 auto E_D1 = 0.04 * CM;
-                auto delD1 = 0.0125 * pow(CW[i], 0.5);
-                auto var_1 = pow(delD1, 2) / 12.;
-                auto sigma1 = sqrt(log(var_1 / pow(E_D1, 2) + 1));
-                auto mu1 = log(E_D1) - pow(sigma1, 2) / 2.;
[…]
