### 2022-10-24 1973c0b
avoid using kw args in equilibrate_wet_radii so that SzyszkowskiLangmuir sigma can be used
