### 2022-10-24 e01d1e9
replace NaN with a proper T value
