### 2022-10-25 26475a7
wip on GPU breakup: fragmentation functions (#936)
Co-authored-by: Sylwester Arabas <sylwester.arabas@uj.edu.pl>
Co-authored-by: Oleksii Bulenok <oleksii.bulenok@student.uj.edu.pl>
