### 2022-10-26 fe34210
refactor fragmentation functions, move straub_p1 to Formulae
