+++ PySDM/physics/fragmentation_function/feingold1988frag.py
@@ -1,0 +1,8 @@
+"""
+Formulae supporting `PySDM.dynamics.collisions.breakup_fragmentations.feingold1988`
+"""
+
+
+class Feingold1988Frag:  # pylint: disable=too-few-public-methods
+    def __init__(self, _):
+        pass
