### 2022-10-26 26743ba
bump test-time-requirements; remove redundant assert
