### 2022-10-26 fc69617
moving Straub p2 into Formulae
