+++ tests/unit_tests/backends/test_collisions_methods.py
@@ -10,7 +10,6 @@ from PySDM.backends.impl_common.indexed_storage import make_IndexedStorage
 from PySDM.backends.impl_common.pair_indicator import make_PairIndicator
 from PySDM.backends.impl_numba.methods.collisions_methods import (
     pair_indices,
-    straub_p2,
     straub_p3,
     straub_p4,
 )
@@ -186,21 +185,6 @@ class TestCollisionMethods:
                 )
         np.testing.assert_array_almost_equal(_gamma.to_ndarray(), expected_gamma)
         np.testing.assert_array_equal(_n_substep, np.asarray(expected_n_substep))
-
-    @staticmethod
-    def test_straub_p2(backend_class=CPU):  # pylint: disable=redefined-outer-name
-        # arrange
-        backend = backend_class()
-        i = 0
-        cw_data = backend.Storage.from_ndarray(np.asarray([0.666])).data
-        frag_size = backend.Storage.from_ndarray(np.asarray([0.0]))
-        rand_data = backend.Storage.from_ndarray(np.asarray([0])).data
-
-        # act
-        straub_p2(i=i, CW=cw_data, frag_size=frag_size.data, rand=rand_data)
-
-        # assert
-        np.testing.assert_approx_equal(frag_size.to_ndarray(), 4.3000510e-09)
 
     @staticmethod
     def test_straub_p3(backend_class=CPU):  # pylint: disable=redefined-outer-name
