+++ PySDM/physics/constants_defaults.py
@@ -162,3 +162,4 @@ ABIFM_UNIT = 1 / si.cm**2 / si.s
 J_HET = np.nan
 
 STRAUB_E_D1 = 0.04 * si.cm
+STRAUB_MU2 = 0.095 * si.cm
