### 2022-10-26 e11e168
moving Straub p3 into Formulae
