+++ tests/unit_tests/backends/test_collisions_methods.py
@@ -8,11 +8,7 @@ from PySDM.backends import CPU
 from PySDM.backends.impl_common.index import make_Index
 from PySDM.backends.impl_common.indexed_storage import make_IndexedStorage
 from PySDM.backends.impl_common.pair_indicator import make_PairIndicator
-from PySDM.backends.impl_numba.methods.collisions_methods import (
-    pair_indices,
-    straub_p3,
-    straub_p4,
-)
+from PySDM.backends.impl_numba.methods.collisions_methods import pair_indices, straub_p4
 
 from ...backends_fixture import backend_class
 
@@ -185,22 +181,6 @@ class TestCollisionMethods:
                 )
         np.testing.assert_array_almost_equal(_gamma.to_ndarray(), expected_gamma)
         np.testing.assert_array_equal(_n_substep, np.asarray(expected_n_substep))
-
-    @staticmethod
-    def test_straub_p3(backend_class=CPU):  # pylint: disable=redefined-outer-name
-        # arrange
-        backend = backend_class()
-        i = 0
-        cw_data = backend.Storage.from_ndarray(np.asarray([0.666])).data
-        ds_data = backend.Storage.from_ndarray(np.asarray([0.0])).data
-        frag_size = backend.Storage.from_ndarray(np.asarray([0.0]))
-        rand_data = backend.Storage.from_ndarray(np.asarray([0])).data
-
-        # act
-        straub_p3(i=i, CW=cw_data, ds=ds_data, frag_size=frag_size.data, rand=rand_data)
-
-        # assert
-        np.testing.assert_approx_equal(frag_size.to_ndarray(), 1.3857897e-15)
 
     @staticmethod
     def test_straub_p4(backend_class=CPU):  # pylint: disable=redefined-outer-name
