### 2022-10-26 7afe65e
moving Straub p4 into Formulae
