+++ tests/unit_tests/backends/test_collisions_methods.py
@@ -4,11 +4,10 @@ import os
 import numpy as np
 import pytest
 
-from PySDM.backends import CPU
 from PySDM.backends.impl_common.index import make_Index
 from PySDM.backends.impl_common.indexed_storage import make_IndexedStorage
 from PySDM.backends.impl_common.pair_indicator import make_PairIndicator
-from PySDM.backends.impl_numba.methods.collisions_methods import pair_indices, straub_p4
+from PySDM.backends.impl_numba.methods.collisions_methods import pair_indices
 
 from ...backends_fixture import backend_class
 
@@ -181,31 +180,3 @@ class TestCollisionMethods:
                 )
         np.testing.assert_array_almost_equal(_gamma.to_ndarray(), expected_gamma)
         np.testing.assert_array_equal(_n_substep, np.asarray(expected_n_substep))
-
-    @staticmethod
-    def test_straub_p4(backend_class=CPU):  # pylint: disable=redefined-outer-name
-        # arrange
-        backend = backend_class()
-        i = 0
-        cw_data = backend.Storage.from_ndarray(np.asarray([0.666])).data
-        ds_data = backend.Storage.from_ndarray(np.asarray([0.0])).data
-        frag_size = backend.Storage.from_ndarray(np.asarray([0.0]))
-        v_max_data = backend.Storage.from_ndarray(np.asarray([0])).data
-        nr1_data = backend.Storage.from_ndarray(np.asarray([1])).data
-        nr2_data = backend.Storage.from_ndarray(np.asarray([2])).data
-        nr3_data = backend.Storage.from_ndarray(np.asarray([0])).data
-
-        # act
-        straub_p4(
-            i=i,
-            CW=cw_data,
-            ds=ds_data,
-            v_max=v_max_data,
-            frag_size=frag_size.data,
-            Nr1=nr1_data,
-            Nr2=nr2_data,
-            Nr3=nr3_data,
-        )
-
-        # assert
-        np.testing.assert_approx_equal(frag_size.to_ndarray(), -5.6454883153e-06)
