+++ PySDM/physics/fragmentation_function/straub2010nf.py
@@ -64,3 +64,33 @@ class Straub2010Nf:  # pylint: disable=too-few-public-methods
             )
             ** 3
         )
+
+    @staticmethod
+    def p4(const, CW, ds, v_max, Nr1, Nr2, Nr3):  # pylint: disable=too-many-arguments
+        return (
+            const.PI
+            / 6
+            * (
+                v_max / const.PI_4_3 * 8
+                + ds**3
+                - Nr1
+                * np.exp(
+                    3 * np.log(const.STRAUB_E_D1)
+                    + 6
+                    * np.log(
+                        (0.0125 * CW**0.5) ** 2 / 12 / const.STRAUB_E_D1**2 + 1
+                    )
+                    / 2
+                )
+                - Nr2
+                * (
+                    const.STRAUB_MU2**3
+                    + 3 * const.STRAUB_MU2 * ((0.007 * (CW - 21.0)) ** 2 / 12) ** 2
+                )
+                - Nr3
+                * (
+                    (0.9 * ds) ** 3
+                    + 3 * 0.9 * ds * ((0.01 * (0.76 * CW**0.5 + 1.0)) ** 2 / 12) ** 2
+                )
+            )
+        )
