### 2022-10-26 d68a9fc
moving Gaussian frag func to Formulae
