### 2022-10-28 9df1813
Extract flag_zero_multiplicity logic for GPU into Commons C struct
