### 2022-10-28 cb342c3
Extract flag_zero_multiplicity logic for GPU into Commons C struct
