### 2022-10-28 8ce8d1c
Remove atomicAdd which is not ready
