### 2022-10-31 40be73c
Remove newline characters when parsing cpp atomicAdd
