+++ PySDM/backends/impl_thrust_rtc/test_helpers/cpp2python.py
@@ -115,6 +115,7 @@ def replace_atomic_mins(cpp: str) -> (str, bool):
 def atomic_add_to_python(cpp: str) -> str:
     cpp = (
         cpp.replace("atomicAdd", "")
+        .replace("\n", "")
         .replace("unsigned long long int*", "")
         .replace("unsigned long long int", "")
         .replace("double*", "")
