### 2022-10-31 06c4c7b
Extract coalesce into Commons Cpp struct and add atomicAdd for compute gamma and coalesce Cpp implementations
