### 2022-10-31 ac3b8a7
Remove unused ith_coalescence_rate variable
