### 2022-10-31 4389e28
code clarifications in Formulae ctor
