### 2022-10-31 b05edba
Add smoke test on Bieli moments
