### 2022-10-31 91343a1
Asserts in bieli smoke test
