### 2022-11-01 363ead2
Add smoke test on Bieli mass conservation (#950)
Co-authored-by: Sylwester Arabas <sylwester.arabas@uj.edu.pl>
