### 2022-11-01 9b21072
add pylint check for max file length
