### 2022-11-02 307a307
Fix to pairwise_methods and potential for zero multiplicity when rounding
