### 2022-11-02 bc51ced
add a draft of Fig_3 smoke test for dejong_and_mackay_2022
