### 2022-11-03 c8ccf8a
add unit test for `find_pairs`
