### 2022-11-03 098d756
add max(N, 1) rounding logic in `break_up_while`
