### 2022-11-04 d8f36e4
reduce resolution and introduce asserts in breakup smoke test (Fig 3)
