### 2022-11-04 9df911f
adapt test_rate_product to updated rate products logic
