### 2022-11-04 d160639
Create test_fig_4.py
