+++ tests/smoke_tests/dejong_and_mackay_2022/test_fig_4.py
@@ -1,0 +1,155 @@
+# pylint: disable=missing-module-docstring,missing-class-docstring,missing-function-docstring
+import matplotlib
+import numpy as np
+from matplotlib import pyplot
+from PySDM_examples.deJong_Mackay_2022 import Settings0D, run_box_breakup
+
+from PySDM.dynamics.collisions.breakup_fragmentations import AlwaysN, Gaussian
+from PySDM.dynamics.collisions.coalescence_efficiencies import ConstEc
+from PySDM.physics.constants import si
+
+CMAP = matplotlib.cm.get_cmap("viridis")
+N_SD = 2**12
+DT = 1 * si.s
+BINS_EDGES = np.logspace(
+    np.log10(5e0 * si.um), np.log10(5e3 * si.um), num=64, endpoint=True
+)
+
+
+class TestFig4:
+    @staticmethod
+    def test_fig_4a(plot=False):
+        # arrange
+        settings = Settings0D()
+        settings.coal_eff = ConstEc(Ec=0.95)
+        settings.n_sd = N_SD
+        settings.radius_bins_edges = BINS_EDGES
+        settings.warn_overflows = False
+        settings.dt = DT
+        nf_vals = [1, 4, 16, 64]
+        data_x = {}
+        data_y = {}
+
+        # act
+        lbl = "initial"
+        (data_x[lbl], data_y[lbl], _) = run_box_breakup(settings, [0])
+        for (i, nf_val) in enumerate(nf_vals):
+            settings.fragmentation = AlwaysN(n=nf_val)
+            lbl = "n_f = " + str(nf_val)
+            (data_x[lbl], data_y[lbl], _) = run_box_breakup(settings, [120])
+
+        # plot
+        pyplot.step(
+            data_x["initial"],
+            data_y["initial"][0] * settings.rho,
+            color="k",
+            linestyle="--",
+            label="initial",
+        )
+        for (i, nf_val) in enumerate(nf_vals):
+            lbl = "n_f = " + str(nf_val)
+            pyplot.step(
+                data_x[lbl],
+                data_y[lbl][0] * settings.rho,
+                color=CMAP(i / len(nf_vals)),
+                label=lbl
+                if lbl not in pyplot.gca().get_legend_handles_labels()[1]
+                else "",
+            )
+        pyplot.xscale("log")
+        pyplot.xlabel("particle radius (um)")
+        pyplot.ylabel("dm/dlnR (kg/m$^3$ / unit(ln R)")
+        pyplot.legend()
+        if plot:
+            pyplot.show()
+
+        # assert
+        for datum_x in data_x.values():
+            np.testing.assert_array_equal(data_x["initial"], datum_x)
+
+        peaks_expected = {
+            "initial": (30, 0.017),
+            "n_f = 1": (1600, 0.015),
+            "n_f = 4": (500, 0.01),
+            "n_f = 16": (200, 0.0075),
+            "n_f = 64": (24, 0.006),
+        }
+
+        for lbl, x_y in peaks_expected.items():
+            print(lbl)
+            peak = np.argmax(data_y[lbl][0])
+            np.testing.assert_approx_equal(data_x[lbl][peak], x_y[0], significant=1)
+            np.testing.assert_approx_equal(
+                data_y[lbl][0][peak] * settings.rho, x_y[1], significant=1
+            )
+
+    @staticmethod
+    def test_fig_4b(plot=False):
+        # arrange
+        settings = Settings0D()
+        settings.dt = DT
+        settings.n_sd = N_SD
+        settings.warn_overflows = False
+        settings.radius_bins_edges = BINS_EDGES
+        x_0 = settings.X0
[…]
