### 2022-11-04 a375f44
take rhod into account in test_rate_product
