### 2022-11-05 0a637ed
fix random seed in test_fig_4a
