### 2022-11-05 58e28ca
add Fig_3 and Fig_4 smoke tests for dejong_and_mackay_2022 (#954)
Co-authored-by: edejong-caltech <edejong@caltech.edu>
Co-authored-by: Oleksii Bulenok <63914185+abulenok@users.noreply.github.com>
