### 2022-11-06 f65ef72
add smoke test for deJong Fig 5
