### 2022-11-08 653ad88
Remove obsolete xfail (test passes)
