### 2022-11-08 80c81d0
separate return statement from attribute-update code block
