### 2022-11-11 0258c38
Streamline arithmetics
