### 2022-11-11 ec38f65
refactor fragmentation functions, move `straub_*` to `Formulae` (#2)
* refactor fragmentation functions, move straub_p1 to Formulae

* bump test-time-requirements; remove redundant assert

* moving Straub p2 into Formulae

* moving Straub p3 into Formulae

* moving Straub p4 into Formulae

* moving Gaussian frag func to Formulae

* moving Feingold1988 frag func to Formulae

* bump test-time-requirements for PySDM-examples

* code clarifications in Formulae ctor
