+++ PySDM/formulae.py
@@ -40,7 +40,27 @@ class Formulae:  # pylint: disable=too-few-public-methods,too-many-instance-attr
         hydrostatics: str = "Default",
         freezing_temperature_spectrum: str = "Null",
         heterogeneous_ice_nucleation_rate: str = "Null",
+        fragmentation_function: str = "AlwaysN",
     ):
+        # initialisation of the fields below is just to silence pylint and to enable code hints
+        # in PyCharm and alike, all these fields are later overwritten within this ctor
+        self.condensation_coordinate = condensation_coordinate
+        self.saturation_vapour_pressure = saturation_vapour_pressure
+        self.hygroscopicity = hygroscopicity
+        self.drop_growth = drop_growth
+        self.surface_tension = surface_tension
+        self.diffusion_kinetics = diffusion_kinetics
+        self.latent_heat = latent_heat
+        self.diffusion_thermics = diffusion_thermics
+        self.ventilation = ventilation
+        self.state_variable_triplet = state_variable_triplet
+        self.particle_advection = particle_advection
+        self.hydrostatics = hydrostatics
+        self.freezing_temperature_spectrum = freezing_temperature_spectrum
+        self.heterogeneous_ice_nucleation_rate = heterogeneous_ice_nucleation_rate
+        self.fragmentation_function = fragmentation_function
+        components = tuple(i for i in dir(self) if not i.startswith("__"))
+
         constants_defaults = {
             k: getattr(defaults, k)
             for defaults in (physics.constants, physics.constants_defaults)
@@ -59,92 +79,19 @@ class Formulae:  # pylint: disable=too-few-public-methods,too-many-instance-attr
             "Trivia", physics.trivia, fastmath, constants, dimensional_analysis
         )
 
-        self.condensation_coordinate = _magick(
-            condensation_coordinate,
-            physics.condensation_coordinate,
-            fastmath,
-            constants,
-            dimensional_analysis,
-        )
-        self.saturation_vapour_pressure = _magick(
-            saturation_vapour_pressure,
-            physics.saturation_vapour_pressure,
-            fastmath,
-            constants,
-            dimensional_analysis,
-        )
-        self.latent_heat = _magick(
-            latent_heat, physics.latent_heat, fastmath, constants, dimensional_analysis
-        )
-        self.hygroscopicity = _magick(
-            hygroscopicity,
-            physics.hygroscopicity,
-            fastmath,
-            constants,
-            dimensional_analysis,
-        )
-        self.drop_growth = _magick(
-            drop_growth, physics.drop_growth, fastmath, constants, dimensional_analysis
-        )
-        self.surface_tension = _magick(
-            surface_tension,
-            physics.surface_tension,
-            fastmath,
-            constants,
-            dimensional_analysis,
-        )
-        self.diffusion_kinetics = _magick(
-            diffusion_kinetics,
-            physics.diffusion_kinetics,
-            fastmath,
-            constants,
-            dimensional_analysis,
-        )
-        self.diffusion_thermics = _magick(
-            diffusion_thermics,
-            physics.diffusion_thermics,
-            fastmath,
-            constants,
-            dimensional_analysis,
-        )
-        self.ventilation = _magick(
-            ventilation, physics.ventilation, fastmath, constants, dimensional_analysis
-        )
-        self.state_variable_triplet = _magick(
-            state_variable_triplet,
-            physics.state_variable_triplet,
-            fastmath,
-            constants,
-            dimensional_analysis,
-        )
-        self.particle_advection = _magick(
-            particle_advection,
-            physics.particle_advection,
-            fastmath,
-            constants,
-            dimensional_analysis,
-        )
-        self.hydrostatics = _magick(
-            hydrostatics,
-            physics.hydrostatics,
-            fastmath,
-            constants,
[…]
