+++ tests/unit_tests/dynamics/collisions/test_sdm_breakup.py
@@ -3,7 +3,7 @@ import numpy as np
 import pytest
 
 import PySDM.physics.constants as const
-from PySDM import Builder
+from PySDM import Builder, Formulae
 from PySDM.backends import CPU
 from PySDM.backends.impl_common.pair_indicator import make_PairIndicator
 from PySDM.dynamics import Breakup
@@ -43,7 +43,9 @@ class TestSDMBreakup:
         nsteps = 10
 
         n_sd = len(attributes["n"])
-        builder = Builder(n_sd, backend_class())
+        builder = Builder(
+            n_sd, backend_class(Formulae(fragmentation_function="AlwaysN"))
+        )
         builder.set_environment(Box(dv=1 * si.cm**3, dt=dt))
         builder.add_dynamic(breakup)
         particulator = builder.build(attributes=attributes, products=())

[…]
