### 2022-11-11 8597e13
Move handle_all_breakups flag from Collision Dynamic to Formulae
This is done enable selection between break_up and break_up_while implementations
