### 2022-11-11 ce6e04d
Pass Formulae to SettingsOD in deJong_Mackay smoke tests
