### 2022-11-16 1e1728f
Add absrtract classes for storages and tests for implementation with Numba
