+++ PySDM/storages/common/storage.py
@@ -1,0 +1,239 @@
+import abc
+from collections.abc import Sequence
+from typing import Any, NamedTuple, Optional, Sized, Tuple, Type, Union
+
+import numpy as np
+from typing_extensions import TypeGuard
+
+from PySDM.storages.common.random import Random
+
+ShapeType = Union[int, Tuple[int, ...]]
+
+
+class StorageSignature(NamedTuple):
+    """groups items defining a storage"""
+
+    data: np.ndarray
+    shape: ShapeType
+    dtype: Type
+
+
+_OtherType = Union["Storage", np.ndarray]
+
+
+class MathMagicMethodsMixin:
+    def __pow__(self, other: _OtherType):
+        raise TypeError("Use **=")
+
+    @abc.abstractmethod
+    def __ipow__(self, other: _OtherType):
+        raise NotImplementedError()
+
+    def __mod__(self, other: _OtherType):
+        raise TypeError("Use %=")
+
+    @abc.abstractmethod
+    def __imod__(self, other: _OtherType):
+        raise NotImplementedError()
+
+    def __truediv__(self, other: _OtherType):
+        raise TypeError("Use /=")
+
+    @abc.abstractmethod
+    def __itruediv__(self, other: _OtherType):
+        raise NotImplementedError()
+
+    def __mul__(self, other: _OtherType):
+        raise TypeError("Use *=")
+
+    @abc.abstractmethod
+    def __imul__(self, other: _OtherType):
+        raise NotImplementedError()
+
+    def __sub__(self, other: _OtherType):
+        raise TypeError("Use -=")
+
+    @abc.abstractmethod
+    def __isub__(self, other: _OtherType):
+        raise NotImplementedError()
+
+    def __add__(self, other: _OtherType):
+        raise TypeError("Use +=")
+
+    @abc.abstractmethod
+    def __iadd__(self, other: _OtherType):
+        raise NotImplementedError()
+
+
+class StorageOperationsMixin(Sequence):
+    @abc.abstractmethod
+    def to_ndarray(self, *args, **kwargs) -> np.ndarray:
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def upload(self, data: np.ndarray):
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def download(self, target: np.ndarray, reshape: bool = False):
+        raise NotImplementedError()
+
+    @classmethod
+    @abc.abstractmethod
+    def _get_empty_data(cls, shape: ShapeType, dtype: Type) -> StorageSignature:
+        raise NotImplementedError()
+
+    @classmethod
+    @abc.abstractmethod
+    def _get_data_from_ndarray(cls, array: np.ndarray) -> StorageSignature:
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def amin(self):
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def floor(self, other: Optional[_OtherType] = None) -> "Storage":
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def product(self, multiplicand: _OtherType, multiplier: _OtherType) -> "Storage":
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def ratio(self, dividend: _OtherType, divisor: _OtherType) -> "Storage":
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def divide_if_not_zero(self, divisor: _OtherType) -> "Storage":
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def sum(self, arg_a: _OtherType, arg_b: _OtherType) -> "Storage":
+        raise NotImplementedError()
+
+    @abc.abstractmethod
[…]
