+++ PySDM/storages/numba/backend/pair.py
@@ -1,0 +1,158 @@
+"""
+CPU implementation of pairwise operations backend methods
+"""
+import numba
+import numpy as np
+
+from PySDM.backends.impl_numba import conf
+from PySDM.storages.common import backend
+
+
+class PairBackend(backend.PairBackend):
+    @staticmethod
+    @numba.njit(**conf.JIT_FLAGS)
+    def distance_pair_body(data_out, data_in, is_first_in_pair, idx, length):
+        data_out[:] = 0
+        for i in numba.prange(length - 1):  # pylint: disable=not-an-iterable
+            if is_first_in_pair[i]:
+                data_out[i // 2] = np.abs(data_in[idx[i]] - data_in[idx[i + 1]])
+
+    @staticmethod
+    def distance_pair(data_out, data_in, is_first_in_pair, idx):
+        return PairBackend.distance_pair_body(
+            data_out.data,
+            data_in.data,
+            is_first_in_pair.indicator.data,
+            idx.data,
+            len(idx),
+        )
+
+    @staticmethod
+    @numba.njit(**conf.JIT_FLAGS)
+    def find_pairs_body(
+        *, cell_start, is_first_in_pair, cell_id, cell_idx, idx, length
+    ):
+        for i in numba.prange(length - 1):  # pylint: disable=not-an-iterable
+            is_first_in_pair[i] = (
+                cell_id[idx[i]] == cell_id[idx[i + 1]]
+                and (i - cell_start[cell_idx[cell_id[idx[i]]]]) % 2 == 0
+            )
+        is_first_in_pair[-1] = False
+
+    @staticmethod
+    def find_pairs(cell_start, is_first_in_pair, cell_id, cell_idx, idx):
+        return PairBackend.find_pairs_body(
+            cell_start=cell_start.data,
+            is_first_in_pair=is_first_in_pair.indicator.data,
+            cell_id=cell_id.data,
+            cell_idx=cell_idx.data,
+            idx=idx.data,
+            length=len(idx),
+        )
+
+    @staticmethod
+    @numba.njit(**conf.JIT_FLAGS)
+    def max_pair_body(data_out, data_in, is_first_in_pair, idx, length):
+        data_out[:] = 0
+        for i in numba.prange(length - 1):  # pylint: disable=not-an-iterable
+            if is_first_in_pair[i]:
+                data_out[i // 2] = max(data_in[idx[i]], data_in[idx[i + 1]])
+
+    @staticmethod
+    def max_pair(data_out, data_in, is_first_in_pair, idx):
+        return PairBackend.max_pair_body(
+            data_out.data,
+            data_in.data,
+            is_first_in_pair.indicator.data,
+            idx.data,
+            len(idx),
+        )
+
+    @staticmethod
+    @numba.njit(**conf.JIT_FLAGS)
+    def min_pair_body(data_out, data_in, is_first_in_pair, idx, length):
+        data_out[:] = 0
+        for i in numba.prange(length - 1):  # pylint: disable=not-an-iterable
+            if is_first_in_pair[i]:
+                data_out[i // 2] = min(data_in[idx[i]], data_in[idx[i + 1]])
+
+    @staticmethod
+    def min_pair(data_out, data_in, is_first_in_pair, idx):
+        return PairBackend.min_pair_body(
+            data_out.data,
+            data_in.data,
+            is_first_in_pair.indicator.data,
+            idx.data,
+            len(idx),
+        )
+
+    @staticmethod
+    @numba.njit(**conf.JIT_FLAGS)
+    def sort_pair_body(data_out, data_in, is_first_in_pair, idx, length):
+        data_out[:] = 0
[…]
