### 2022-11-17 d8e0503
Add some more tests for index
