### 2022-11-17 #4412 @claresinger
@dhhagan can you please try again? I think the correct command is `add` not `invite`
