### 2022-11-17 56a57ef
Extract common logic for break_up and break_up_while
