### 2022-11-22 ac60c12
Fix typings and add some tests
