### 2022-11-23 887adf9
Add test for indexed storage
