### 2022-11-23 0493f38
Finish testing indexed storage
