### 2022-11-23 964398d
Add Pair Indicator
