### 2022-11-25 c75b347
Fix pylint and pre-commit errors
