+++ tests/smoke_tests/bieli_et_al_2022/test_moments.py
@@ -6,12 +6,13 @@ from matplotlib import pyplot
 from PySDM_examples.Bieli_et_al_2022.settings import Settings
 from PySDM_examples.Bieli_et_al_2022.simulation import make_core
 
+from PySDM import Formulae
 from PySDM.physics import si
 
 
 def test_moments(plot=False):
     # arrange
-    settings = Settings(Formulae(fragentation_function='Feingold1988Frag'))
+    settings = Settings(Formulae(fragentation_function="Feingold1988Frag"))
     if "CI" in os.environ:
         settings.n_sd = 10
     else:
