### 2022-11-25 429dbcf
Improve variable naming for break_up
