### 2022-11-25 #4968 @editorialbot
[REVIEW]: New developments in PySDM and PySDM-examples v2: collisional breakup, immersion freezing, dry aerosol composition initialisation, and adaptive time-stepping
**Submitting author:** @edejong-caltech (<a href="http://orcid.org/0000-0002-5310-4554">Emily de Jong</a>)
**Repository:** https://github.com/atmos-cloud-sim-uj/PySDM
**Branch with paper.md** (empty if default branch):
**Version:** v2.8
