### 2022-11-25 #4412 @slayoo
@dhhagan

> `@editorialbot add @claresinger as reviewer`

Clare is an author here, not a reviewer.
The intended command likely was: `@editorialbot add @emmasimp as reviewer`
