+++ tests/smoke_tests/dejong_and_mackay_2022/test_fig_4.py
@@ -4,7 +4,6 @@ import numpy as np
 from matplotlib import pyplot
 from PySDM_examples.deJong_Mackay_2022 import Settings0D, run_box_breakup
 
-from PySDM import Formulae
 from PySDM.dynamics.collisions.breakup_fragmentations import AlwaysN, Gaussian
 from PySDM.dynamics.collisions.coalescence_efficiencies import ConstEc
 from PySDM.physics.constants import si
