### 2022-11-26 add33e3
reverting plot=True setting in test
