### 2022-11-28 9cc3f7e
breakup refactors (#966)
Co-authored-by: Sylwester Arabas <sylwester.arabas@uj.edu.pl>

- fragmentation functions partially moved into Formulae
- breakup backend methods split into smaller functions that are common across `breakup` and `breakup_while`
- `handle_all_breakups` flag added to Formulae
