+++ PySDM/backends/impl_thrust_rtc/methods/collisions_methods.py
@@ -28,7 +28,6 @@ class CollisionsMethods(
 ):  # pylint: disable=too-many-instance-attributes
     def __init__(self):
         ThrustRTCBackendMethods.__init__(self)
-        const = self.formulae.constants
 
         self.__adaptive_sdm_gamma_body_1 = trtc.For(
             ("dt_todo", "dt_left", "dt_max"),
@@ -249,15 +248,20 @@ class CollisionsMethods(
             """,
         )
 
-        self.__gauss_fragmentation_body = trtc.For(
-            param_names=("mu", "sigma", "frag_size", "rand"),
-            name_iter="i",
-            body=f"""
-            frag_size[i] = mu - sigma / {const.sqrt_two} / {const.sqrt_pi} / log(2.) * log(
-                (0.5 + rand[i]) / (1.5 - rand[i])
-            );
-            """,
-        )
+        if self.formulae.fragmentation_function.__name__ == "Gaussian":
+            self.__gauss_fragmentation_body = trtc.For(
+                param_names=("mu", "sigma", "frag_size", "rand"),
+                name_iter="i",
+                body=f"""
+                frag_size[i] = {self.formulae.fragmentation_function.frag_size.c_inline(
+                    mu="mu",
+                    sigma="sigma",
+                    rand="rand[i]"
+                )};
+                """.replace(
+                    "real_type", self._get_c_type()
+                ),
+            )
 
         self.__slams_fragmentation_body = trtc.For(
             param_names=("n_fragment", "frag_size", "x_plus_y", "probs", "rand"),
@@ -277,141 +281,92 @@ class CollisionsMethods(
             """,
         )
 
-        self.__feingold1988_fragmentation_body = trtc.For(
-            param_names=("scale", "frag_size", "x_plus_y", "rand", "fragtol"),
-            name_iter="i",
-            body="""
-            auto log_arg = max(1 - rand[i] * scale / x_plus_y[i], fragtol);
-            frag_size[i] = -scale * log(log_arg);
-            """,
-        )
+        if self.formulae.fragmentation_function.__name__ == "Feingold1988Frag":
+            self.__feingold1988_fragmentation_body = trtc.For(
+                param_names=("scale", "frag_size", "x_plus_y", "rand", "fragtol"),
+                name_iter="i",
+                body=f"""
+                frag_size[i] = {self.formulae.fragmentation_function.frag_size.c_inline(
+                    scale="scale",
+                    rand="rand[i]",
+                    x_plus_y="x_plus_y[i]",
+                    fragtol="fragtol"
+                )};
+                """.replace(
+                    "real_type", self._get_c_type()
+                ),
+            )
 
         self.__straub_Nr_body = """
-
-            if (gam[i] * CW[i] >= 7.0) {
-                Nr1[i] = 0.088 * (gam[i] * CW[i] - 7.0);
-            }
-            if (CW[i] >= 21.0) {
-                Nr2[i] = 0.22 * (CW[i] - 21.0);
-                if (CW[i] <= 46.0) {
-                    Nr3[i] = 0.04 * (46.0 - CW[i]);
+                if (gam[i] * CW[i] >= 7.0) {
+                    Nr1[i] = 0.088 * (gam[i] * CW[i] - 7.0);
                 }
-            }
-            else {
-                Nr3[i] = 1.0;
-            }
-            Nr4[i] = 1.0;
-            Nrt[i] = Nr1[i] + Nr2[i] + Nr3[i] + Nr4[i];
+                if (CW[i] >= 21.0) {
[…]
