+++ PySDM/backends/impl_numba/methods/collisions_methods.py
@@ -10,9 +10,6 @@ from PySDM.backends.impl_numba import conf
 from PySDM.backends.impl_numba.atomic_operations import atomic_add
 from PySDM.backends.impl_numba.storage import Storage
 from PySDM.backends.impl_numba.warnings import warn
-from PySDM.physics.constants import PI, PI_4_3, si, sqrt_pi, sqrt_two
-
-CM = si.cm
 
 
 @numba.njit(**{**conf.JIT_FLAGS, **{"parallel": False}})
@@ -52,68 +49,61 @@ def coalesce(  # pylint: disable=too-many-arguments
 
 
 @numba.njit(**{**conf.JIT_FLAGS, **{"parallel": False}})
-def break_up(  # pylint: disable=too-many-arguments,unused-argument,too-many-locals
-    i,
-    j,
-    k,
-    cid,
-    multiplicity,
-    gamma,
-    attributes,
-    n_fragment,
-    fragment_size,
-    max_multiplicity,
-    breakup_rate,
-    breakup_rate_deficit,
-    warn_overflows,
-    volume,
-):  # pylint: disable=too-many-branches
+def breakup_fun0(
+    gamma, j, k, multiplicity, volume, nfi, fragment_size_i, max_multiplicity
+):  # pylint: disable=too-many-arguments
     overflow_flag = False
+    gamma_j_k = 0
     take_from_j_test = multiplicity[k]
+    take_from_j = 0
     new_mult_k_test = 0
     new_mult_k = multiplicity[k]
-    take_from_j = 0
-    gamma_tmp = 0
-    gamma_deficit = gamma[i]
-    for m in range(int(gamma[i])):
-        take_from_j_test = new_mult_k_test + take_from_j_test
-        new_mult_k_test = (
-            new_mult_k_test * (volume[j] / fragment_size[i])
-            + n_fragment[i] * multiplicity[k]
-        )
+    for m in range(int(gamma)):
+        take_from_j_test += new_mult_k_test
+        new_mult_k_test *= volume[j] / fragment_size_i
+        new_mult_k_test += nfi * multiplicity[k]
+
         # check for overflow of multiplicity
         if new_mult_k_test > max_multiplicity:
             overflow_flag = True
             break
+
         # check for new_n > 0
         if take_from_j_test > multiplicity[j]:
             break
 
-        # all tests passed
         take_from_j = take_from_j_test
         new_mult_k = new_mult_k_test
-        gamma_tmp = m + 1
-        gamma_deficit = gamma[i] - gamma_tmp
-    # 2. Compute the new multiplicities and particle sizes, with rounding
+        gamma_j_k = m + 1
+    return take_from_j, new_mult_k, gamma_j_k, overflow_flag
+
+
+@numba.njit(**{**conf.JIT_FLAGS, **{"parallel": False}})
+def breakup_fun1(
+    j, k, attributes, multiplicity, take_from_j, new_mult_k
+):  # pylint: disable=too-many-arguments
     for a in range(0, len(attributes)):
         attributes[a, k] *= multiplicity[k]
         attributes[a, k] += take_from_j * attributes[a, j]
         attributes[a, k] /= new_mult_k
+
     if multiplicity[j] > take_from_j:
         nj = multiplicity[j] - take_from_j
         nk = new_mult_k
     else:
         nj = new_mult_k / 2
-        if round(nj) == 0:
-            atomic_add(breakup_rate_deficit, cid, gamma[i] * multiplicity[k])
-            return
         nk = nj
+    return nj, nk
+
+
+@numba.njit(**{**conf.JIT_FLAGS, **{"parallel": False}})
+def breakup_fun2(
[…]
