### 2022-11-28 970ce6c
add missing double support for float
