### 2022-11-28 84a1d9e
Rm redundant assignment to n_fragment[i]
