+++ PySDM/backends/impl_thrust_rtc/methods/collisions_methods.py
@@ -242,7 +242,6 @@ class CollisionsMethods(
             }
             if (frag_size[i] == 0.0) {
                 frag_size[i] = x_plus_y[i];
-                n_fragment[i] = 1.0;
             }
             n_fragment[i] = x_plus_y[i] / frag_size[i];
             """,
