### 2022-12-02 d664a88
Add test for Coalesce deficit counters on GPU
