### 2022-12-02 d366801
Add test for Coalesce deficit counters on GPU
