+++ tests/unit_tests/products/test_collision_rates.py
@@ -19,6 +19,10 @@ from PySDM.products import (
     CollisionRatePerGridbox,
 )
 
+from ...backends_fixture import backend_class
+
+assert hasattr(backend_class, "_pytestfixturefunction")
+
 
 class TestCollisionProducts:
     @staticmethod
@@ -128,6 +132,7 @@ class TestCollisionProducts:
         "n_init",
         [[5, 2], [1, 2, 3, 4], [3, 7] * 10],
     )
+    # pylint: disable=redefined-outer-name
     def test_no_collision_deficits_when_adaptive(params, n_init, backend_class=CPU):
         # Arrange
         n_sd = len(n_init)
@@ -173,6 +178,7 @@ class TestCollisionProducts:
             },
         ],
     )
+    # pylint: disable=redefined-outer-name
     def test_breakup_deficits_when_adaptive(params, backend_class=CPU):
         # Arrange
         n_init = [7, 353]
@@ -220,6 +226,7 @@ class TestCollisionProducts:
             },
         ],
     )
+    # pylint: disable=redefined-outer-name
     def test_no_breakup_deficits_when_while_loop(params, backend_class=CPU):
         # Arrange
         n_init = [7, 353]
@@ -274,6 +281,7 @@ class TestCollisionProducts:
             },
         ],
     )
+    # pylint: disable=redefined-outer-name
     def test_rate_sums_single_cell(params, backend_class=CPU):
         # Arrange
         n_init = [7, 353]
