### 2022-12-02 6ff8f89
Add pair_indices function to GPU
