### 2022-12-02 e55728d
Commons::coalesce in GPU backend incl. support for deficit counters (#949)
Co-authored-by: Sylwester Arabas <sylwester.arabas@uj.edu.pl>
