+++ PySDM/dynamics/collisions/breakup_efficiencies/constEb.py
@@ -12,4 +12,4 @@ class ConstEb:
         self.particulator = builder.particulator
 
     def __call__(self, output, is_first_in_pair):
-        output.data[:] = self.Eb
+        output[:] = self.Eb
