### 2022-12-04 17d9443
Add a test for pair indicator
