### 2022-12-04 cedb2cc
Fix base classes
