### 2022-12-04 ed25766
Add tests for pairwise storage
