### 2022-12-04 387aa90
Add a test for min in pairwise
