+++ PySDM/storages/numba/tests/test_pairwise.py
@@ -41,23 +41,23 @@ def indexed_storage_cls():
 
 
 @pytest.fixture()
-def pair_indicator_and_pairwise_storage(
-    index_backend, pair_backend, pair_indicator_cls, pairwise_storage_cls, index_cls
-):
+def pair_indicator_instance(pair_indicator_cls):
     # Arrange
-    pair_indicator_instance = pair_indicator_cls(6)
+    pair_indicator_instance_ = pair_indicator_cls(6)
     pair_indicator_data = np.ones(6)
     pair_indicator_data[np.arange(1, 6, 2)] = 0
-    pair_indicator_instance.indicator = pair_indicator_data
-    pairwise_storage_instance = pairwise_storage_cls.empty(3, Storage.FLOAT)
-    return pair_indicator_instance, pairwise_storage_instance
+    pair_indicator_instance_.indicator = pair_indicator_data
+    return pair_indicator_instance_
 
 
-def test_distance(pair_indicator_and_pairwise_storage, indexed_storage_cls, index_cls):
-    (
-        pair_indicator_instance,
-        pairwise_storage_instance,
-    ) = pair_indicator_and_pairwise_storage
+@pytest.fixture()
+def pairwise_storage_instance(pairwise_storage_cls):
+    return pairwise_storage_cls.empty(3, Storage.FLOAT)
+
+
+def test_distance(
+    pair_indicator_instance, pairwise_storage_instance, indexed_storage_cls, index_cls
+):
     data = np.asarray([1, 4, 2, 5, 3, 6])
     data_idx = index_cls.identity_index(6)
     pairwise_storage_instance.distance(
@@ -67,11 +67,9 @@ def test_distance(pair_indicator_and_pairwise_storage, indexed_storage_cls, inde
     np.testing.assert_allclose(pairwise_storage_instance.data, np.asarray([3, 3, 3]))
 
 
-def test_max(pair_indicator_and_pairwise_storage, indexed_storage_cls, index_cls):
-    (
-        pair_indicator_instance,
-        pairwise_storage_instance,
-    ) = pair_indicator_and_pairwise_storage
+def test_max(
+    pair_indicator_instance, pairwise_storage_instance, indexed_storage_cls, index_cls
+):
     data = np.asarray([1, -1, 2, 4, 3, 3])
     data_idx = index_cls.identity_index(6)
     pairwise_storage_instance.max(
@@ -81,11 +79,9 @@ def test_max(pair_indicator_and_pairwise_storage, indexed_storage_cls, index_cls
     np.testing.assert_allclose(pairwise_storage_instance.data, np.asarray([1, 4, 3]))
 
 
-def test_min(pair_indicator_and_pairwise_storage, indexed_storage_cls, index_cls):
-    (
-        pair_indicator_instance,
-        pairwise_storage_instance,
-    ) = pair_indicator_and_pairwise_storage
+def test_min(
+    pair_indicator_instance, pairwise_storage_instance, indexed_storage_cls, index_cls
+):
     data = np.asarray([1, -1, 2, 4, 3, 3])
     data_idx = index_cls.identity_index(6)
     pairwise_storage_instance.min(
@@ -93,3 +89,43 @@ def test_min(pair_indicator_and_pairwise_storage, indexed_storage_cls, index_cls
         pair_indicator_instance,
     )
     np.testing.assert_allclose(pairwise_storage_instance.data, np.asarray([-1, 2, 3]))
+
+
+def test_multiply(
+    pair_indicator_instance, pairwise_storage_instance, indexed_storage_cls, index_cls
+):
+    data = np.asarray([1, -1, 2, 4, 3, 3])
[…]
