### 2022-12-08 3b07198
common code for computing concentrations at STP
