+++ PySDM/products/impl/standard_temperature_pressure_concentration_product.py
@@ -1,0 +1,37 @@
+"""
+common code for products computing particle concentrations with
+the option to return them at standard temperature and pressure conditions
+"""
+from abc import ABC
+
+from PySDM.products.impl.moment_product import MomentProduct
+
+
+class StandardTemperaturePressureConcentrationProduct(ABC, MomentProduct):
+    def __init__(self, *, unit: str, name: str, specific: bool, stp: bool):
+        super().__init__(unit=unit, name=name)
+        self.specific = specific
+        self.stp = stp
+        self.rho_stp = None
+        if self.stp and self.specific:
+            raise ValueError(
+                "std-temperature-and-pressure precludes specific conc. option"
+            )
+
+    def register(self, builder):
+        super().register(builder)
+        self.rho_stp = builder.formulae.constants.rho_STP
+
+    def _impl(self):
+        self.buffer[:] /= self.particulator.mesh.dv
+
+        if self.specific or self.stp:
+            result = self.buffer.copy()
+            self._download_to_buffer(self.particulator.environment["rhod"])
+            result[:] /= self.buffer
+            if self.stp:
+                result[:] *= self.rho_stp
+        else:
+            result = self.buffer
+
+        return result
