+++ PySDM/products/size_spectral/particle_concentration.py
@@ -4,16 +4,25 @@ concentration of particles within a grid cell (either per-volume of per-mass-of-
 """
 import numpy as np
 
-from PySDM.products.impl.moment_product import MomentProduct
+from PySDM.products.impl.standard_temperature_pressure_concentration_product import (
+    StandardTemperaturePressureConcentrationProduct,
+)
 
 
-class ParticleConcentration(MomentProduct):
+class ParticleConcentration(StandardTemperaturePressureConcentrationProduct):
     def __init__(
-        self, radius_range=(0, np.inf), specific=False, name=None, unit="m^-3"
+        self,
+        radius_range=(0, np.inf),
+        specific=False,
+        stp=False,
+        name=None,
+        unit="m^-3",
     ):
         self.radius_range = radius_range
-        self.specific = specific
-        super().__init__(name=name, unit=unit)
+        super().__init__(name=name, unit=unit, specific=specific, stp=stp)
+
+    def register(self, builder):
+        super().register(builder)
 
     def _impl(self, **kwargs):
         self._download_moment_to_buffer(
@@ -24,14 +33,7 @@ class ParticleConcentration(MomentProduct):
                 self.formulae.trivia.volume(self.radius_range[1]),
             ),
         )
-        self.buffer[:] /= self.particulator.mesh.dv
-        if self.specific:
-            result = self.buffer.copy()  # TODO #217
-            self._download_to_buffer(self.particulator.environment["rhod"])
-            result[:] /= self.buffer
-        else:
-            result = self.buffer
-        return result
+        return super()._impl()
 
 
 class ParticleSpecificConcentration(ParticleConcentration):
