### 2022-12-08 0d8dcc0
Add "out" argument to compute_gamma and compute_probabilities_of_collision
"out" is a location into which the result is stored
