+++ PySDM/backends/impl_thrust_rtc/methods/collisions_methods.py
@@ -55,9 +55,9 @@ struct Commons {
       int64_t i,
       VectorView<int64_t> idx,
       VectorView<bool> is_first_in_pair,
-      VectorView<real_type> gamma
+      VectorView<real_type> prob_or_gamma
   ) {
-      if (gamma[i] == 0) {
+      if (prob_or_gamma[i] == 0) {
           return std::make_tuple(-1, -1, true);
       }
       auto offset = 1 - is_first_in_pair[2 * i];
@@ -84,19 +84,19 @@ class CollisionsMethods(
         )
 
         self.__adaptive_sdm_gamma_body_2 = trtc.For(
-            ("gamma", "idx", "n", "cell_id", "dt", "is_first_in_pair", "dt_todo"),
+            ("prob", "idx", "n", "cell_id", "dt", "is_first_in_pair", "dt_todo"),
             "i",
             f"""
                 {COMMONS}
                 bool skip_pair = false;
                 int64_t j = -1;
                 int64_t k = -1;
-                std::tie(j, k, skip_pair) = Commons::pair_indices(i, idx, is_first_in_pair, gamma);
+                std::tie(j, k, skip_pair) = Commons::pair_indices(i, idx, is_first_in_pair, prob);
                 if (skip_pair) {{
                     return;
                 }}
                 auto prop = (int64_t)(n[j] / n[k]);
-                auto dt_optimal = dt * prop / gamma[i];
+                auto dt_optimal = dt * prop / prob[i];
                 auto cid = cell_id[j];
                 static_assert(sizeof(dt_todo[0]) == sizeof(unsigned int), "");
                 atomicMin((unsigned int*)&dt_todo[cid], __float_as_uint(dt_optimal));
@@ -106,18 +106,18 @@ class CollisionsMethods(
         )
 
         self.__adaptive_sdm_gamma_body_3 = trtc.For(
-            ("gamma", "idx", "cell_id", "dt", "is_first_in_pair", "dt_todo"),
+            ("prob", "idx", "cell_id", "dt", "is_first_in_pair", "dt_todo", "out"),
             "i",
             f"""
                 {COMMONS}
                 bool skip_pair = false;
                 int64_t j = -1;
                 int64_t k = -1;
-                std::tie(j, k, skip_pair) = Commons::pair_indices(i, idx, is_first_in_pair, gamma);
+                std::tie(j, k, skip_pair) = Commons::pair_indices(i, idx, is_first_in_pair, prob);
                 if (skip_pair) {{
                     return;
                 }}
-                gamma[i] *= dt_todo[cell_id[j]] / dt;
+                out[i] *= dt_todo[cell_id[j]] / dt;
             """.replace(
                 "real_type", self._get_c_type()
             ),
@@ -182,7 +182,7 @@ class CollisionsMethods(
                 "real_type", self._get_c_type()
             ),
         )
-
+        # pylint:disable=unused-private-member
         self.__collision_coalescence_breakup_body = trtc.For(
             param_names=(
                 "multiplicity",
@@ -227,29 +227,31 @@ class CollisionsMethods(
 
         self.__compute_gamma_body = trtc.For(
             (
-                "gamma",
+                "prob",
                 "rand",
                 "idx",
                 "multiplicity",
                 "cell_id",
                 "collision_rate_deficit",
                 "collision_rate",
+                "is_first_in_pair",
+                "out",
             ),
             "i",
             f"""
             {COMMONS}
-            gamma[i] = ceil(gamma[i] - rand[i]);
+            out[i] = ceil(prob[i] - rand[i]);
 
             bool skip_pair = false;
             int64_t j = -1;
             int64_t k = -1;
-            std::tie(j, k, skip_pair) = Commons::pair_indices(i, idx, is_first_in_pair, gamma);
[…]
