### 2022-12-08 6eb0888
Add missing arguments to some collision functions
