### 2022-12-09 2157857
Add missing arguments in adaptive_sdm_gamma tests
