### 2022-12-09 7fa3051
refactors (thanks to pylint)
