+++ PySDM/products/size_spectral/particle_concentration.py
@@ -4,9 +4,7 @@ concentration of particles within a grid cell (either per-volume of per-mass-of-
 """
 import numpy as np
 
-from PySDM.products.impl.standard_temperature_pressure_concentration_product import (
-    ConcentrationProduct,
-)
+from PySDM.products.impl.concentration_product import ConcentrationProduct
 
 
 class ParticleConcentration(ConcentrationProduct):
