### 2023-01-31 #4968 @douglowe
Apologies for the length of time it has taken for this review.

After looking over the code and paper, I'm happy to report that it looks to be in great shape to me. The paper is well written and describes the new functionality well, the API documentation is easy to find and useful, and the test coverage of the code seems good.

The one thing I would note is that there is not any discussion of the 'state of the field' in this paper. The version 1 paper does list other implementations of these routines, which are also noted in the README document for the code repository. But it would be useful here to have a short note on what extra (if any) open source packages have been added since v1, and if any other software tools have overlapping functionality specifically with the new features described in this paper.
