### 2023-02-03 #4968 @emmasimp
Sorry for the delay in getting this review done!

Overall I think the paper and code are very well written. I found the code and documentation easy to understand and get up and running with the model. I wish more software was documented in a way that allowed me to install and run the model within a matter of minutes!

I agree with Doug that the only thing missing is a bit more detail in the paper of how the new features of PySDM v2 fit in to the wider picture of similar models available (or whether or not any exist).
