### 2023-02-13 #4412 @douglowe
Does the article proof need regenerating in the review thread rather than here? https://github.com/openjournals/joss-reviews/issues/4968
