### 2023-02-14 #4968 @slayoo
Thank you!

One question for the editors we have is how to technically handle the fact that the project has recently moved from `github.com/atmos-cloud-sim-uj` to `github.com/open-atmos`. There is a redirect, so the previous link does work, but it would be great to promote the new location on the JOSS website next to the paper. @dhhagan @Kevin-Mattheus-Moerman is there a way to change such piece of metadata? Thanks
