### 2023-03-06 #4968 @Kevin-Mattheus-Moerman
After you've completed the above ☝️ , Please work on the following:

- [ ] Please archive a copy of the software on ZENODO. You can do this manually, or you may follow these automated steps: https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content
- [ ] Once you have the above archive, please report back here with the archive DOI.
- [ ] Please ensure that the ZENODO archive has the same title and author list (and order) as the paper, and the same license (it might list "Other(open)" by default) as your software. You may need to manually edit these aspects.
- [ ] Please let me know if `v2.8` is still the current version tag for the JOSS publication. It may have altered during review or be incremented during archiving. Note that the version listed on ZENODO, should match the version listed here, and should also be a release tag on your repository.
