### 2023-03-07 #4968 @slayoo
> @edejong-caltech @slayoo I have checked the paper just now. All looks good but can you please edit your affiliations so that USA is spelled out as United States of America?

Voilà: https://github.com/open-atmos/PySDM/commit/0463f3040d25ac3dbad46de744212f31b1c0f13b

@Kevin-Mattheus-Moerman One question concerning Zenodo metadata: in the repo, we maintain the `https://github.com/open-atmos/PySDM/blob/main/.zenodo.json` file where we list project contributors. These include developers who contributed to the v1 releases and were part of the JOSS v1 paper author list (https://joss.theoj.org/papers/10.21105/joss.03219) as well as developers who contributed to the v2 releases who are on the v2 paper author list (i.e., this review). How to meet the "ensure that the ZENODO archive has the same ... author list (and order) as the paper" requirement in such case?

Concerning version tag, I suggest we use the current release, i.e. v2.16:
- https://github.com/open-atmos/PySDM/releases/tag/v2.16
- https://doi.org/10.5281/zenodo.7640495
