### 2023-03-14 #4968 @Kevin-Mattheus-Moerman
@edejong-caltech @slayoo Would it be okay to make the ZENODO archive author set match that of the current paper? Also can you update the title to match the paper title too? You can manually edit these.
If you can process the author update then there is no issue.
