### 2023-03-22 #4968 @slayoo
> If you can process the author update then there is no issue.

Zenodo metadata updated: https://zenodo.org/record/7640495
